+++ cypress/e2e/bulk-edit/csv/bulk-edit-csv-users-address-type.cy.js
@@ -8,6 +8,7 @@ import Users from '../../../support/fragments/users/users';
 import UsersSearchPane from '../../../support/fragments/users/usersSearchPane';
 import UserEdit from '../../../support/fragments/users/userEdit';
 import ExportFile from '../../../support/fragments/data-export/exportFile';
+import BulkEditLogs from '../../../support/fragments/bulk-edit/bulk-edit-logs';
 
 let user;
 let addressTypeId;
@@ -101,21 +102,21 @@ describe('bulk-edit', () => {
                 BulkEditActions.downloadMatchedRecordsAbsent();
 
                 BulkEditSearchPane.openLogsSearch();
-                BulkEditSearchPane.verifyLogsPane();
-                BulkEditSearchPane.checkUsersCheckbox();
-                BulkEditSearchPane.clickActionsRunBy(user.username);
-                BulkEditSearchPane.verifyLogsRowActionWhenCompleted();
+                BulkEditLogs.verifyLogsPane();
+                BulkEditLogs.checkUsersCheckbox();
+                BulkEditLogs.clickActionsRunBy(user.username);
+                BulkEditLogs.verifyLogsRowActionWhenCompleted();
 
-                BulkEditSearchPane.downloadFileUsedToTrigger();
+                BulkEditLogs.downloadFileUsedToTrigger();
                 ExportFile.verifyFileIncludes(userBarcodesFileName, [user.barcode]);
 
-                BulkEditSearchPane.downloadFileWithMatchingRecords();
+                BulkEditLogs.downloadFileWithMatchingRecords();
                 ExportFile.verifyFileIncludes(matchedRecordsFileName, [user.barcode]);
 
-                BulkEditSearchPane.downloadFileWithProposedChanges();
+                BulkEditLogs.downloadFileWithProposedChanges();
                 ExportFile.verifyFileIncludes(previewOfProposedChangesFileName, [newFirstName]);
 
-                BulkEditSearchPane.downloadFileWithUpdatedRecords();
+                BulkEditLogs.downloadFileWithUpdatedRecords();
                 ExportFile.verifyFileIncludes(updatedRecordsFileName, [newFirstName]);
 
                 cy.visit(TopMenu.usersPath);
