+++ cypress/e2e/bulk-edit/in-app/bulk-edit-in-app-user-custom-fields-semicolon.cy.js
@@ -11,6 +11,7 @@ import Users from '../../../support/fragments/users/users';
 import UsersSearchPane from '../../../support/fragments/users/usersSearchPane';
 import FileManager from '../../../support/utils/fileManager';
 import getRandomPostfix from '../../../support/utils/stringTools';
+import BulkEditLogs from '../../../support/fragments/bulk-edit/bulk-edit-logs';
 
 // TO DO: remove ignoring errors. Now when you click on one of the buttons, some promise in the application returns false
 Cypress.on('uncaught:exception', () => false);
@@ -145,13 +146,13 @@ describe('bulk-edit', () => {
         ]);
 
         BulkEditSearchPane.openLogsSearch();
-        BulkEditSearchPane.checkUsersCheckbox();
-        BulkEditSearchPane.clickActionsRunBy(secondUser.username);
-        BulkEditSearchPane.downloadFileWithProposedChanges();
+        BulkEditLogs.checkUsersCheckbox();
+        BulkEditLogs.clickActionsRunBy(secondUser.username);
+        BulkEditLogs.downloadFileWithProposedChanges();
         ExportFile.verifyFileIncludes(previewOfProposedChangesFileName, [
           `${customFieldData.fieldLabel}:${customFieldData.label1};${customFieldData.label2}`,
         ]);
-        BulkEditSearchPane.downloadFileWithUpdatedRecords();
+        BulkEditLogs.downloadFileWithUpdatedRecords();
         ExportFile.verifyFileIncludes(changedRecordsFileName, [
           `${customFieldData.fieldLabel}:${customFieldData.label1};${customFieldData.label2}`,
         ]);
