+++ cypress/e2e/bulk-edit/in-app/bulk-edit-in-app-user-toggles.cy.js
@@ -6,6 +6,7 @@ import Users from '../../../support/fragments/users/users';
 import DateTools from '../../../support/utils/dateTools';
 import FileManager from '../../../support/utils/fileManager';
 import getRandomPostfix from '../../../support/utils/stringTools';
+import BulkEditLogs from '../../../support/fragments/bulk-edit/bulk-edit-logs';
 
 let user;
 const userUUIDsFileName = `userUUIDs_${getRandomPostfix()}.csv`;
@@ -59,21 +60,21 @@ describe('bulk-edit', () => {
           'Completed with errors',
           'Failed',
         ];
-        statuses.forEach((status) => BulkEditSearchPane.checkLogsCheckbox(status));
+        statuses.forEach((status) => BulkEditLogs.checkLogsCheckbox(status));
 
-        BulkEditSearchPane.checkHoldingsCheckbox();
-        BulkEditSearchPane.checkUsersCheckbox();
-        BulkEditSearchPane.checkItemsCheckbox();
-        BulkEditSearchPane.fillLogsStartDate(today, today);
-        BulkEditSearchPane.applyStartDateFilters();
-        BulkEditSearchPane.fillLogsEndDate(today, today);
-        BulkEditSearchPane.applyEndDateFilters();
-        BulkEditSearchPane.verifyLogResultsFound();
+        BulkEditLogs.checkHoldingsCheckbox();
+        BulkEditLogs.checkUsersCheckbox();
+        BulkEditLogs.checkItemsCheckbox();
+        BulkEditLogs.fillLogsStartDate(today, today);
+        BulkEditLogs.applyStartDateFilters();
+        BulkEditLogs.fillLogsEndDate(today, today);
+        BulkEditLogs.applyEndDateFilters();
+        BulkEditLogs.verifyLogResultsFound();
         BulkEditSearchPane.openIdentifierSearch();
         BulkEditSearchPane.verifyMatchedResults(user.username);
         BulkEditSearchPane.openLogsSearch();
-        BulkEditSearchPane.resetAll();
-        BulkEditSearchPane.verifyLogsPane();
+        BulkEditLogs.resetAll();
+        BulkEditLogs.verifyLogsPane();
         BulkEditSearchPane.openIdentifierSearch();
         BulkEditSearchPane.verifyMatchedResults(user.username);
         BulkEditSearchPane.openIdentifierSearch();
@@ -86,20 +87,20 @@ describe('bulk-edit', () => {
         BulkEditSearchPane.waitFileUploading();
         BulkEditSearchPane.verifyErrorLabelAfterChanges(userUUIDsFileName, 0, 1);
         BulkEditSearchPane.openLogsSearch();
-        BulkEditSearchPane.verifyLogsPane();
-        statuses.forEach((status) => BulkEditSearchPane.checkLogsCheckbox(status));
-        BulkEditSearchPane.checkHoldingsCheckbox();
-        BulkEditSearchPane.checkUsersCheckbox();
-        BulkEditSearchPane.checkItemsCheckbox();
-        BulkEditSearchPane.fillLogsStartDate(today, today);
-        BulkEditSearchPane.applyStartDateFilters();
-        BulkEditSearchPane.fillLogsEndDate(today, today);
-        BulkEditSearchPane.applyEndDateFilters();
-        BulkEditSearchPane.verifyLogResultsFound();
+        BulkEditLogs.verifyLogsPane();
+        statuses.forEach((status) => BulkEditLogs.checkLogsCheckbox(status));
+        BulkEditLogs.checkHoldingsCheckbox();
+        BulkEditLogs.checkUsersCheckbox();
+        BulkEditLogs.checkItemsCheckbox();
+        BulkEditLogs.fillLogsStartDate(today, today);
+        BulkEditLogs.applyStartDateFilters();
+        BulkEditLogs.fillLogsEndDate(today, today);
+        BulkEditLogs.applyEndDateFilters();
+        BulkEditLogs.verifyLogResultsFound();
         BulkEditSearchPane.openIdentifierSearch();
         BulkEditSearchPane.verifyErrorLabelAfterChanges(userUUIDsFileName, 0, 1);
         BulkEditSearchPane.openLogsSearch();
-        BulkEditSearchPane.resetAll();
+        BulkEditLogs.resetAll();
         BulkEditSearchPane.openIdentifierSearch();
         BulkEditSearchPane.verifyErrorLabelAfterChanges(userUUIDsFileName, 0, 1);
       },
