+++ cypress/e2e/bulk-edit/in-app/bulk-edit-in-app-verify-default-state.cy.js
@@ -57,7 +57,7 @@ describe('bulk-edit', () => {
         BulkEditSearchPane.verifySpecificTabHighlighted('Identifier');
         BulkEditSearchPane.verifyRecordIdentifierEmpty();
         BulkEditSearchPane.isDragAndDropAreaDisabled(true);
-        BulkEditSearchPane.verifyRecordTypesSortedAlphabetically(false);
+        BulkEditSearchPane.verifyRecordTypesSortedAlphabetically();
 
         BulkEditSearchPane.checkHoldingsRadio();
         BulkEditSearchPane.isHoldingsRadioChecked(true);
