+++ cypress/e2e/bulk-edit/logs/bulk-edit-logs-table-view.cy.js
@@ -3,47 +3,50 @@ import BulkEditSearchPane from '../../../support/fragments/bulk-edit/bulk-edit-s
 import TopMenu from '../../../support/fragments/topMenu';
 import Users from '../../../support/fragments/users/users';
 import DateTools from '../../../support/utils/dateTools';
+import BulkEditLogs from '../../../support/fragments/bulk-edit/bulk-edit-logs';
 
 let user;
 
-describe('Bulk Edit - Logs', () => {
-  before('create test data', () => {
-    cy.createTempUser([
-      permissions.bulkEditLogsView.gui,
-      permissions.bulkEditCsvView.gui,
-      permissions.bulkEditView.gui,
-    ]).then((userProperties) => {
-      user = userProperties;
-      cy.login(user.username, user.password, {
-        path: TopMenu.bulkEditPath,
-        waiter: BulkEditSearchPane.waitLoading,
+describe('bulk-edit', () => {
+  describe('logs', () => {
+    before('create test data', () => {
+      cy.createTempUser([
+        permissions.bulkEditLogsView.gui,
+        permissions.bulkEditCsvView.gui,
+        permissions.bulkEditView.gui,
+      ]).then((userProperties) => {
+        user = userProperties;
+        cy.login(user.username, user.password, {
+          path: TopMenu.bulkEditPath,
+          waiter: BulkEditSearchPane.waitLoading,
+        });
       });
     });
+
+    after('delete test data', () => {
+      cy.getAdminToken();
+      Users.deleteViaApi(user.userId);
+    });
+
+    it(
+      'C368015 Verify that displays a table in the main logs page (firebird)',
+      { tags: ['smoke', 'firebird'] },
+      () => {
+        const tomorrowDate = DateTools.getFormattedDate(
+          { date: DateTools.getTomorrowDay() },
+          'YYYY-MM-DD',
+        );
+
+        BulkEditSearchPane.openLogsSearch();
+        BulkEditLogs.verifyLogsPane();
+        BulkEditLogs.checkHoldingsCheckbox();
+        BulkEditLogs.checkUsersCheckbox();
+        BulkEditLogs.checkItemsCheckbox();
+        BulkEditLogs.verifyLogsTableHeaders();
+        BulkEditLogs.fillLogsStartDate(tomorrowDate, tomorrowDate);
+        BulkEditLogs.applyStartDateFilters();
+        BulkEditLogs.noLogResultsFound();
+      },
+    );
   });
-
-  after('delete test data', () => {
-    cy.getAdminToken();
-    Users.deleteViaApi(user.userId);
-  });
-
-  it(
-    'C368015 Verify that displays a table in the main logs page (firebird)',
-    { tags: ['smoke', 'firebird'] },
-    () => {
-      const tomorrowDate = DateTools.getFormattedDate(
-        { date: DateTools.getTomorrowDay() },
-        'YYYY-MM-DD',
-      );
-
-      BulkEditSearchPane.openLogsSearch();
-      BulkEditSearchPane.verifyLogsPane();
-      BulkEditSearchPane.checkHoldingsCheckbox();
-      BulkEditSearchPane.checkUsersCheckbox();
-      BulkEditSearchPane.checkItemsCheckbox();
-      BulkEditSearchPane.verifyLogsTableHeaders();
-      BulkEditSearchPane.fillLogsStartDate(tomorrowDate, tomorrowDate);
-      BulkEditSearchPane.applyStartDateFilters();
-      BulkEditSearchPane.noLogResultsFound();
-    },
-  );
 });
