+++ cypress/e2e/bulk-edit/logs/csv/bulk-edit-logs-csv-users-barcodes.cy.js
@@ -6,6 +6,7 @@ import getRandomPostfix from '../../../../support/utils/stringTools';
 import FileManager from '../../../../support/utils/fileManager';
 import BulkEditActions from '../../../../support/fragments/bulk-edit/bulk-edit-actions';
 import BulkEditFiles from '../../../../support/fragments/bulk-edit/bulk-edit-files';
+import BulkEditLogs from '../../../../support/fragments/bulk-edit/bulk-edit-logs';
 
 let user;
 let userWithoutPermissions;
@@ -22,126 +23,130 @@ const previewOfProposedChangesFileName = {
 const updatedRecordsFileName = `*-Changed-Records*-${invalidAndValidUserBarcodesFileName}`;
 const errorsFromCommittingFileName = `*-Committing-changes-Errors-${invalidAndValidUserBarcodesFileName}`;
 
-describe('Bulk Edit - Logs', () => {
-  before('create test data', () => {
-    cy.createTempUser([]).then((userProperties) => {
-      userWithoutPermissions = userProperties;
-    });
-    cy.createTempUser([
-      permissions.bulkEditLogsView.gui,
-      permissions.bulkEditCsvView.gui,
-      permissions.bulkEditCsvEdit.gui,
-      permissions.uiUserEdit.gui,
-    ]).then((userProperties) => {
-      user = userProperties;
-      cy.login(user.username, user.password, {
-        path: TopMenu.bulkEditPath,
-        waiter: BulkEditSearchPane.waitLoading,
+describe('bulk-edit', () => {
+  describe('logs', () => {
+    describe('csv approach', () => {
+      before('create test data', () => {
+        cy.createTempUser([]).then((userProperties) => {
+          userWithoutPermissions = userProperties;
+        });
+        cy.createTempUser([
+          permissions.bulkEditLogsView.gui,
+          permissions.bulkEditCsvView.gui,
+          permissions.bulkEditCsvEdit.gui,
+          permissions.uiUserEdit.gui,
+        ]).then((userProperties) => {
+          user = userProperties;
+          cy.login(user.username, user.password, {
+            path: TopMenu.bulkEditPath,
+            waiter: BulkEditSearchPane.waitLoading,
+          });
+          FileManager.createFile(
+            `cypress/fixtures/${invalidAndValidUserBarcodesFileName}`,
+            `${user.barcode}\n${userWithoutPermissions.barcode}\n${invalidUserBarcode}`,
+          );
+        });
       });
-      FileManager.createFile(
-        `cypress/fixtures/${invalidAndValidUserBarcodesFileName}`,
-        `${user.barcode}\n${userWithoutPermissions.barcode}\n${invalidUserBarcode}`,
+
+      after('delete test data', () => {
+        cy.getAdminToken();
+        FileManager.deleteFile(`cypress/fixtures/${invalidAndValidUserBarcodesFileName}`);
+        FileManager.deleteFile(`cypress/fixtures/${editedFileName}`);
+        Users.deleteViaApi(user.userId);
+        Users.deleteViaApi(userWithoutPermissions.userId);
+        FileManager.deleteFileFromDownloadsByMask(
+          invalidAndValidUserBarcodesFileName,
+          `*${matchedRecordsFileName}`,
+          previewOfProposedChangesFileName.first,
+          updatedRecordsFileName,
+          errorsFromCommittingFileName,
+          errorsFromMatchingFileName,
+        );
+      });
+
+      it(
+        'C375215 Verify generated Logs files for Users CSV - with errors (firebird)',
+        { tags: ['smoke', 'firebird'] },
+        () => {
+          BulkEditSearchPane.verifyDragNDropRecordTypeIdentifierArea('Users', 'User Barcodes');
+          BulkEditSearchPane.uploadFile(invalidAndValidUserBarcodesFileName);
+          BulkEditSearchPane.waitFileUploading();
+
+          BulkEditActions.downloadMatchedResults();
+          BulkEditActions.downloadErrors();
+
+          BulkEditActions.prepareValidBulkEditFile(
+            matchedRecordsFileName,
+            editedFileName,
+            user.firstName,
+            newFirstName,
+          );
+
+          BulkEditActions.openStartBulkEditForm();
+          BulkEditSearchPane.uploadFile(editedFileName);
+          BulkEditSearchPane.waitFileUploading();
+          BulkEditActions.clickNext();
+          BulkEditActions.commitChanges();
+
+          BulkEditSearchPane.verifyChangedResults(newFirstName);
+          BulkEditActions.openActions();
+          BulkEditActions.downloadChangedCSV();
+          BulkEditActions.downloadErrors();
+
+          BulkEditSearchPane.openLogsSearch();
+          BulkEditLogs.verifyLogsPane();
+          BulkEditLogs.checkUsersCheckbox();
+          BulkEditLogs.clickActionsRunBy(user.username);
+          BulkEditLogs.verifyLogsRowActionWhenCompletedWithErrors();
+
+          BulkEditLogs.downloadFileUsedToTrigger();
+          BulkEditFiles.verifyCSVFileRows(invalidAndValidUserBarcodesFileName, [
+            user.barcode,
+            userWithoutPermissions.barcode,
+            invalidUserBarcode,
+          ]);
+
+          BulkEditLogs.downloadFileWithMatchingRecords();
+          BulkEditFiles.verifyMatchedResultFileContent(
+            `*${matchedRecordsFileName}`,
+            [user.barcode, userWithoutPermissions.barcode],
+            'userBarcode',
+            true,
+          );
+
+          BulkEditLogs.downloadFileWithErrorsEncountered();
+          BulkEditFiles.verifyMatchedResultFileContent(
+            errorsFromMatchingFileName,
+            [invalidUserBarcode],
+            'firstElement',
+            false,
+          );
+
+          BulkEditLogs.downloadFileWithProposedChanges();
+          BulkEditFiles.verifyMatchedResultFileContent(
+            previewOfProposedChangesFileName.first,
+            [newFirstName, userWithoutPermissions.firstName],
+            'firstName',
+            true,
+          );
+
+          BulkEditLogs.downloadFileWithUpdatedRecords();
+          BulkEditFiles.verifyMatchedResultFileContent(
+            updatedRecordsFileName,
+            [newFirstName, userWithoutPermissions.firstName],
+            'firstName',
+            true,
+          );
+
+          BulkEditLogs.downloadFileWithCommitErrors();
+          BulkEditFiles.verifyMatchedResultFileContent(
+            errorsFromCommittingFileName,
+            [userWithoutPermissions.barcode],
+            'firstElement',
+            false,
+          );
+        },
       );
     });
   });
-
-  after('delete test data', () => {
-    cy.getAdminToken();
-    FileManager.deleteFile(`cypress/fixtures/${invalidAndValidUserBarcodesFileName}`);
-    FileManager.deleteFile(`cypress/fixtures/${editedFileName}`);
-    Users.deleteViaApi(user.userId);
-    Users.deleteViaApi(userWithoutPermissions.userId);
-    FileManager.deleteFileFromDownloadsByMask(
-      invalidAndValidUserBarcodesFileName,
-      `*${matchedRecordsFileName}`,
-      previewOfProposedChangesFileName.first,
-      updatedRecordsFileName,
-      errorsFromCommittingFileName,
-      errorsFromMatchingFileName,
-    );
-  });
-
-  it(
-    'C375215 Verify generated Logs files for Users CSV - with errors (firebird)',
-    { tags: ['smoke', 'firebird'] },
-    () => {
-      BulkEditSearchPane.verifyDragNDropRecordTypeIdentifierArea('Users', 'User Barcodes');
-      BulkEditSearchPane.uploadFile(invalidAndValidUserBarcodesFileName);
-      BulkEditSearchPane.waitFileUploading();
-
-      BulkEditActions.downloadMatchedResults();
-      BulkEditActions.downloadErrors();
-
-      BulkEditActions.prepareValidBulkEditFile(
-        matchedRecordsFileName,
-        editedFileName,
-        user.firstName,
-        newFirstName,
-      );
-
-      BulkEditActions.openStartBulkEditForm();
-      BulkEditSearchPane.uploadFile(editedFileName);
-      BulkEditSearchPane.waitFileUploading();
-      BulkEditActions.clickNext();
-      BulkEditActions.commitChanges();
-
-      BulkEditSearchPane.verifyChangedResults(newFirstName);
-      BulkEditActions.openActions();
-      BulkEditActions.downloadChangedCSV();
-      BulkEditActions.downloadErrors();
-
-      BulkEditSearchPane.openLogsSearch();
-      BulkEditSearchPane.verifyLogsPane();
-      BulkEditSearchPane.checkUsersCheckbox();
-      BulkEditSearchPane.clickActionsRunBy(user.username);
-      BulkEditSearchPane.verifyLogsRowActionWhenCompletedWithErrors();
-
-      BulkEditSearchPane.downloadFileUsedToTrigger();
-      BulkEditFiles.verifyCSVFileRows(invalidAndValidUserBarcodesFileName, [
-        user.barcode,
-        userWithoutPermissions.barcode,
-        invalidUserBarcode,
-      ]);
-
-      BulkEditSearchPane.downloadFileWithMatchingRecords();
-      BulkEditFiles.verifyMatchedResultFileContent(
-        `*${matchedRecordsFileName}`,
-        [user.barcode, userWithoutPermissions.barcode],
-        'userBarcode',
-        true,
-      );
-
-      BulkEditSearchPane.downloadFileWithErrorsEncountered();
-      BulkEditFiles.verifyMatchedResultFileContent(
-        errorsFromMatchingFileName,
-        [invalidUserBarcode],
-        'firstElement',
-        false,
-      );
-
-      BulkEditSearchPane.downloadFileWithProposedChanges();
-      BulkEditFiles.verifyMatchedResultFileContent(
-        previewOfProposedChangesFileName.first,
-        [newFirstName, userWithoutPermissions.firstName],
-        'firstName',
-        true,
-      );
-
-      BulkEditSearchPane.downloadFileWithUpdatedRecords();
-      BulkEditFiles.verifyMatchedResultFileContent(
-        updatedRecordsFileName,
-        [newFirstName, userWithoutPermissions.firstName],
-        'firstName',
-        true,
-      );
-
-      BulkEditSearchPane.downloadFileWithCommitErrors();
-      BulkEditFiles.verifyMatchedResultFileContent(
-        errorsFromCommittingFileName,
-        [userWithoutPermissions.barcode],
-        'firstElement',
-        false,
-      );
-    },
-  );
 });
