+++ cypress/e2e/bulk-edit/logs/csv/bulk-edit-logs-csv-users-view-permission.cy.js
@@ -8,6 +8,7 @@ import BulkEditActions from '../../../../support/fragments/bulk-edit/bulk-edit-a
 import UsersSearchPane from '../../../../support/fragments/users/usersSearchPane';
 import UserEdit from '../../../../support/fragments/users/userEdit';
 import UsersCard from '../../../../support/fragments/users/usersCard';
+import BulkEditLogs from '../../../../support/fragments/bulk-edit/bulk-edit-logs';
 
 let user;
 const newName = `testName_${getRandomPostfix()}`;
@@ -16,78 +17,85 @@ const matchedRecordsFileName = `Matched-Records-${userUUIDsFileName}`;
 const editedFileName = `edited-records-${getRandomPostfix()}.csv`;
 const changedRecordsFileName = `*-Changed-Records*-${userUUIDsFileName}`;
 
-describe('Bulk Edit - Logs', () => {
-  before('create test data', () => {
-    cy.createTempUser([
-      permissions.bulkEditLogsView.gui,
-      permissions.bulkEditCsvView.gui,
-      permissions.bulkEditCsvEdit.gui,
-      permissions.uiUserEdit.gui,
-    ]).then((userProperties) => {
-      user = userProperties;
-      cy.login(user.username, user.password, {
-        path: TopMenu.bulkEditPath,
-        waiter: BulkEditSearchPane.waitLoading,
+describe('bulk-edit', () => {
+  describe('logs', () => {
+    describe('csv approach', () => {
+      before('create test data', () => {
+        cy.createTempUser([
+          permissions.bulkEditLogsView.gui,
+          permissions.bulkEditCsvView.gui,
+          permissions.bulkEditCsvEdit.gui,
+          permissions.uiUserEdit.gui,
+        ]).then((userProperties) => {
+          user = userProperties;
+          cy.login(user.username, user.password, {
+            path: TopMenu.bulkEditPath,
+            waiter: BulkEditSearchPane.waitLoading,
+          });
+          FileManager.createFile(`cypress/fixtures/${userUUIDsFileName}`, `${user.userId}`);
+        });
       });
-      FileManager.createFile(`cypress/fixtures/${userUUIDsFileName}`, `${user.userId}`);
+
+      after('delete test data', () => {
+        cy.getAdminToken();
+        FileManager.deleteFile(`cypress/fixtures/${userUUIDsFileName}`);
+        FileManager.deleteFile(`cypress/fixtures/${editedFileName}`);
+        Users.deleteViaApi(user.userId);
+        FileManager.deleteFileFromDownloadsByMask(
+          `*${matchedRecordsFileName}`,
+          changedRecordsFileName,
+        );
+      });
+
+      it(
+        'C380562 Verify generated Logs files for Users CSV are hidden without "Users: Can view user profile" permission (firebird)',
+        { tags: ['criticalPath', 'firebird'] },
+        () => {
+          BulkEditSearchPane.verifyDragNDropRecordTypeIdentifierArea('Users', 'User UUIDs');
+          BulkEditSearchPane.uploadFile(userUUIDsFileName);
+          BulkEditSearchPane.waitFileUploading();
+          BulkEditSearchPane.verifyMatchedResults(user.username);
+
+          BulkEditActions.downloadMatchedResults();
+          BulkEditActions.prepareValidBulkEditFile(
+            matchedRecordsFileName,
+            editedFileName,
+            user.firstName,
+            newName,
+          );
+
+          BulkEditActions.openStartBulkEditForm();
+          BulkEditSearchPane.uploadFile(editedFileName);
+          BulkEditSearchPane.waitFileUploading();
+          BulkEditActions.clickNext();
+          BulkEditActions.commitChanges();
+          BulkEditSearchPane.verifyChangedResults(newName);
+
+          BulkEditActions.openActions();
+          BulkEditActions.downloadChangedCSV();
+
+          cy.loginAsAdmin({ path: TopMenu.usersPath, waiter: UsersSearchPane.waitLoading });
+          UsersSearchPane.searchByUsername(user.username);
+          UsersSearchPane.openUser(user.username);
+          UserEdit.addPermissions([permissions.uiUserEdit.gui]);
+          UserEdit.saveAndClose();
+          UsersCard.verifyPermissions([
+            permissions.bulkEditLogsView.gui,
+            permissions.bulkEditCsvView.gui,
+            permissions.bulkEditCsvEdit.gui,
+          ]);
+
+          cy.login(user.username, user.password, {
+            path: TopMenu.bulkEditPath,
+            waiter: BulkEditSearchPane.waitLoading,
+          });
+
+          BulkEditSearchPane.openLogsSearch();
+          BulkEditLogs.verifyLogsPane();
+          BulkEditLogs.checkUsersCheckbox();
+          BulkEditLogs.logActionsIsAbsent();
+        },
+      );
     });
   });
-
-  after('delete test data', () => {
-    cy.getAdminToken();
-    FileManager.deleteFile(`cypress/fixtures/${userUUIDsFileName}`);
-    FileManager.deleteFile(`cypress/fixtures/${editedFileName}`);
-    Users.deleteViaApi(user.userId);
-    FileManager.deleteFileFromDownloadsByMask(`*${matchedRecordsFileName}`, changedRecordsFileName);
-  });
-
-  it(
-    'C380562 Verify generated Logs files for Users CSV are hidden without "Users: Can view user profile" permission (firebird)',
-    { tags: ['criticalPath', 'firebird'] },
-    () => {
-      BulkEditSearchPane.verifyDragNDropRecordTypeIdentifierArea('Users', 'User UUIDs');
-      BulkEditSearchPane.uploadFile(userUUIDsFileName);
-      BulkEditSearchPane.waitFileUploading();
-      BulkEditSearchPane.verifyMatchedResults(user.username);
-
-      BulkEditActions.downloadMatchedResults();
-      BulkEditActions.prepareValidBulkEditFile(
-        matchedRecordsFileName,
-        editedFileName,
-        user.firstName,
-        newName,
-      );
-
-      BulkEditActions.openStartBulkEditForm();
-      BulkEditSearchPane.uploadFile(editedFileName);
-      BulkEditSearchPane.waitFileUploading();
-      BulkEditActions.clickNext();
-      BulkEditActions.commitChanges();
-      BulkEditSearchPane.verifyChangedResults(newName);
-
-      BulkEditActions.openActions();
-      BulkEditActions.downloadChangedCSV();
-
-      cy.loginAsAdmin({ path: TopMenu.usersPath, waiter: UsersSearchPane.waitLoading });
-      UsersSearchPane.searchByUsername(user.username);
-      UsersSearchPane.openUser(user.username);
-      UserEdit.addPermissions([permissions.uiUserEdit.gui]);
-      UserEdit.saveAndClose();
-      UsersCard.verifyPermissions([
-        permissions.bulkEditLogsView.gui,
-        permissions.bulkEditCsvView.gui,
-        permissions.bulkEditCsvEdit.gui,
-      ]);
-
-      cy.login(user.username, user.password, {
-        path: TopMenu.bulkEditPath,
-        waiter: BulkEditSearchPane.waitLoading,
-      });
-
-      BulkEditSearchPane.openLogsSearch();
-      BulkEditSearchPane.verifyLogsPane();
-      BulkEditSearchPane.checkUsersCheckbox();
-      BulkEditSearchPane.logActionsIsAbsent();
-    },
-  );
 });
