+++ cypress/e2e/bulk-edit/logs/csv/bulk-edit-logs-filter-section-users-filter.cy.js
@@ -6,155 +6,164 @@ import TopMenu from '../../../../support/fragments/topMenu';
 import Users from '../../../../support/fragments/users/users';
 import FileManager from '../../../../support/utils/fileManager';
 import getRandomPostfix from '../../../../support/utils/stringTools';
+import BulkEditLogs from '../../../../support/fragments/bulk-edit/bulk-edit-logs';
 
-describe('Bulk Edit - Logs', () => {
-  const users = [];
-  const tastData = {
-    usersAccordion: 'User',
-  };
-  const validItemUUIDsFileName = `validItemUUIDs_${getRandomPostfix()}.csv`;
-  const userBarcodesFileName = `userBarcodes_${getRandomPostfix()}.csv`;
-  const item = {
-    instanceName: `testBulkEdit_${getRandomPostfix()}`,
-    itemBarcode: getRandomPostfix(),
-  };
+describe('bulk-edit', () => {
+  describe('logs', () => {
+    describe('csv approach', () => {
+      const users = [];
+      const tastData = {
+        usersAccordion: 'User',
+      };
+      const validItemUUIDsFileName = `validItemUUIDs_${getRandomPostfix()}.csv`;
+      const userBarcodesFileName = `userBarcodes_${getRandomPostfix()}.csv`;
+      const item = {
+        instanceName: `testBulkEdit_${getRandomPostfix()}`,
+        itemBarcode: getRandomPostfix(),
+      };
 
-  before('Create test data', () => {
-    cy.getAdminToken();
-    cy.createTempUser(
-      [permissions.bulkEditCsvView.gui, permissions.bulkEditLogsView.gui],
-      'faculty',
-    ).then((userProperties) => {
-      users[0] = userProperties;
-    });
-    cy.createTempUser(
-      [permissions.bulkEditView.gui, permissions.bulkEditEdit.gui, permissions.inventoryAll.gui],
-      'faculty',
-    ).then((userProperties) => {
-      users[1] = userProperties;
-    });
-    cy.createTempUser(
-      [permissions.bulkEditUpdateRecords.gui, permissions.uiUserEdit.gui],
-      'faculty',
-    ).then((userProperties) => {
-      users[2] = userProperties;
-    });
-
-    InventoryInstances.createInstanceViaApi(item.instanceName, item.itemBarcode);
-    cy.wait(1000);
-    cy.getInstance({
-      limit: 1,
-      expandAll: true,
-      query: `"items.barcode"=="${item.itemBarcode}"`,
-    })
-      .then((instance) => {
-        item.itemId = instance.items[0].id;
-        FileManager.createFile(`cypress/fixtures/${validItemUUIDsFileName}`, item.itemId);
-        FileManager.createFile(
-          `cypress/fixtures/${userBarcodesFileName}`,
-          `${users[1].barcode}\n${users[2].barcode}`,
-        );
-      })
-      .then(() => {
-        cy.login(users[1].username, users[1].password, {
-          path: TopMenu.bulkEditPath,
-          waiter: BulkEditSearchPane.waitLoading,
+      before('Create test data', () => {
+        cy.getAdminToken();
+        cy.createTempUser(
+          [permissions.bulkEditCsvView.gui, permissions.bulkEditLogsView.gui],
+          'faculty',
+        ).then((userProperties) => {
+          users[0] = userProperties;
         });
-        BulkEditSearchPane.checkItemsRadio();
-        BulkEditSearchPane.selectRecordIdentifier('Item UUIDs');
-        BulkEditSearchPane.uploadFile(validItemUUIDsFileName);
-        BulkEditSearchPane.waitFileUploading();
-
-        cy.login(users[2].username, users[2].password, {
-          path: TopMenu.bulkEditPath,
-          waiter: BulkEditSearchPane.waitLoading,
+        cy.createTempUser(
+          [
+            permissions.bulkEditView.gui,
+            permissions.bulkEditEdit.gui,
+            permissions.inventoryAll.gui,
+          ],
+          'faculty',
+        ).then((userProperties) => {
+          users[1] = userProperties;
         });
-        BulkEditSearchPane.checkUsersRadio();
-        BulkEditSearchPane.selectRecordIdentifier('User Barcodes');
-        BulkEditSearchPane.uploadFile(userBarcodesFileName);
-        BulkEditSearchPane.waitFileUploading();
-
-        BulkEditActions.openActions();
-        BulkEditActions.openInAppStartBulkEditFrom();
-        BulkEditActions.fillPatronGroup('staff (Staff Member)');
-        BulkEditActions.confirmChanges();
-        BulkEditActions.commitChanges();
-        BulkEditSearchPane.waitFileUploading();
-
-        cy.login(users[0].username, users[0].password, {
-          path: TopMenu.bulkEditPath,
-          waiter: BulkEditSearchPane.waitLoading,
+        cy.createTempUser(
+          [permissions.bulkEditUpdateRecords.gui, permissions.uiUserEdit.gui],
+          'faculty',
+        ).then((userProperties) => {
+          users[2] = userProperties;
         });
+
+        InventoryInstances.createInstanceViaApi(item.instanceName, item.itemBarcode);
+        cy.wait(1000);
+        cy.getInstance({
+          limit: 1,
+          expandAll: true,
+          query: `"items.barcode"=="${item.itemBarcode}"`,
+        })
+          .then((instance) => {
+            item.itemId = instance.items[0].id;
+            FileManager.createFile(`cypress/fixtures/${validItemUUIDsFileName}`, item.itemId);
+            FileManager.createFile(
+              `cypress/fixtures/${userBarcodesFileName}`,
+              `${users[1].barcode}\n${users[2].barcode}`,
+            );
+          })
+          .then(() => {
+            cy.login(users[1].username, users[1].password, {
+              path: TopMenu.bulkEditPath,
+              waiter: BulkEditSearchPane.waitLoading,
+            });
+            BulkEditSearchPane.checkItemsRadio();
+            BulkEditSearchPane.selectRecordIdentifier('Item UUIDs');
+            BulkEditSearchPane.uploadFile(validItemUUIDsFileName);
+            BulkEditSearchPane.waitFileUploading();
+
+            cy.login(users[2].username, users[2].password, {
+              path: TopMenu.bulkEditPath,
+              waiter: BulkEditSearchPane.waitLoading,
+            });
+            BulkEditSearchPane.checkUsersRadio();
+            BulkEditSearchPane.selectRecordIdentifier('User Barcodes');
+            BulkEditSearchPane.uploadFile(userBarcodesFileName);
+            BulkEditSearchPane.waitFileUploading();
+
+            BulkEditActions.openActions();
+            BulkEditActions.openInAppStartBulkEditFrom();
+            BulkEditActions.fillPatronGroup('staff (Staff Member)');
+            BulkEditActions.confirmChanges();
+            BulkEditActions.commitChanges();
+            BulkEditSearchPane.waitFileUploading();
+
+            cy.login(users[0].username, users[0].password, {
+              path: TopMenu.bulkEditPath,
+              waiter: BulkEditSearchPane.waitLoading,
+            });
+          });
       });
-  });
 
-  after('delete test data', () => {
-    cy.getAdminToken();
-    InventoryInstances.deleteInstanceAndHoldingRecordAndAllItemsViaApi(item.itemBarcode);
-    users.forEach((user) => {
-      Users.deleteViaApi(user.userId);
+      after('delete test data', () => {
+        cy.getAdminToken();
+        InventoryInstances.deleteInstanceAndHoldingRecordAndAllItemsViaApi(item.itemBarcode);
+        users.forEach((user) => {
+          Users.deleteViaApi(user.userId);
+        });
+        FileManager.deleteFile(`cypress/fixtures/${validItemUUIDsFileName}`);
+        FileManager.deleteFile(`cypress/fixtures/${userBarcodesFileName}`);
+      });
+
+      it(
+        'C409495 Filters section: Users filter (firebird) (TaaS)',
+        { tags: ['extendedPath', 'firebird'] },
+        () => {
+          BulkEditSearchPane.openLogsSearch();
+          BulkEditLogs.verifyLogsPane();
+
+          BulkEditLogs.checkUsersCheckbox();
+          BulkEditLogs.checkHoldingsCheckbox();
+          BulkEditLogs.checkItemsCheckbox();
+          BulkEditLogs.verifyClearSelectedButtonExists('Record types');
+          BulkEditLogs.resetAllBtnIsDisabled(false);
+          BulkEditLogs.verifyLogsTableHeaders();
+
+          BulkEditLogs.clickUserAccordion();
+          BulkEditLogs.clickChooseUserUnderUserAccordion();
+
+          BulkEditLogs.fillUserFilterInput(users[1].username);
+          BulkEditLogs.verifyDropdown(users[1].username);
+          BulkEditLogs.selectUserFromDropdown(users[1].username);
+          BulkEditLogs.verifyClearSelectedButtonExists(tastData.usersAccordion);
+          BulkEditLogs.resetAllBtnIsDisabled(false);
+          BulkEditLogs.verifyCellsValues(
+            8,
+            `${users[1].username}, ${users[1].firstName} ${Users.defaultUser.personal.middleName}`,
+          );
+
+          BulkEditLogs.clickChooseUserUnderUserAccordion();
+          BulkEditLogs.fillUserFilterInput(users[2].username);
+          BulkEditLogs.verifyDropdown(users[2].username);
+          BulkEditLogs.selectUserFromDropdown(users[2].username);
+          BulkEditLogs.verifyClearSelectedButtonExists(tastData.usersAccordion);
+          BulkEditLogs.resetAllBtnIsDisabled(false);
+          BulkEditLogs.verifyCellsValues(
+            8,
+            `${users[2].username}, ${users[2].firstName} ${Users.defaultUser.personal.middleName}`,
+          );
+
+          BulkEditLogs.clickClearSelectedButton(tastData.usersAccordion);
+          BulkEditLogs.verifyClearSelectedButtonExists(tastData.usersAccordion, false);
+          BulkEditLogs.verifyLogsPaneHeader();
+
+          BulkEditLogs.clickChooseUserUnderUserAccordion();
+          BulkEditLogs.fillUserFilterInput(users[1].username);
+          BulkEditLogs.verifyDropdown(users[1].username);
+          BulkEditLogs.selectUserFromDropdown(users[1].username);
+          BulkEditLogs.verifyClearSelectedButtonExists(tastData.usersAccordion);
+          BulkEditLogs.resetAllBtnIsDisabled(false);
+          BulkEditLogs.verifyCellsValues(
+            8,
+            `${users[1].username}, ${users[1].firstName} ${Users.defaultUser.personal.middleName}`,
+          );
+
+          BulkEditLogs.clickChooseUserUnderUserAccordion();
+          BulkEditLogs.fillUserFilterInput(users[0].username);
+          BulkEditLogs.verifyUserIsNotInUserList(users[0].username);
+          BulkEditLogs.verifyEmptyUserDropdown();
+        },
+      );
     });
-    FileManager.deleteFile(`cypress/fixtures/${validItemUUIDsFileName}`);
-    FileManager.deleteFile(`cypress/fixtures/${userBarcodesFileName}`);
   });
-
-  it(
-    'C409495 Filters section: Users filter (firebird) (TaaS)',
-    { tags: ['extendedPath', 'firebird'] },
-    () => {
-      BulkEditSearchPane.openLogsSearch();
-      BulkEditSearchPane.verifyLogsPane();
-
-      BulkEditSearchPane.checkUsersCheckbox();
-      BulkEditSearchPane.checkHoldingsCheckbox();
-      BulkEditSearchPane.checkItemsCheckbox();
-      BulkEditSearchPane.verifyClearSelectedButtonExists('Record types');
-      BulkEditSearchPane.resetAllBtnIsDisabled(false);
-      BulkEditSearchPane.verifyLogsTableHeaders();
-
-      BulkEditSearchPane.clickUserAccordion();
-      BulkEditSearchPane.clickChooseUserUnderUserAccordion();
-
-      BulkEditSearchPane.fillUserFilterInput(users[1].username);
-      BulkEditSearchPane.verifyDropdown(users[1].username);
-      BulkEditSearchPane.selectUserFromDropdown(users[1].username);
-      BulkEditSearchPane.verifyClearSelectedButtonExists(tastData.usersAccordion);
-      BulkEditSearchPane.resetAllBtnIsDisabled(false);
-      BulkEditSearchPane.verifyCellsValues(
-        8,
-        `${users[1].username}, ${users[1].firstName} ${Users.defaultUser.personal.middleName}`,
-      );
-
-      BulkEditSearchPane.clickChooseUserUnderUserAccordion();
-      BulkEditSearchPane.fillUserFilterInput(users[2].username);
-      BulkEditSearchPane.verifyDropdown(users[2].username);
-      BulkEditSearchPane.selectUserFromDropdown(users[2].username);
-      BulkEditSearchPane.verifyClearSelectedButtonExists(tastData.usersAccordion);
-      BulkEditSearchPane.resetAllBtnIsDisabled(false);
-      BulkEditSearchPane.verifyCellsValues(
-        8,
-        `${users[2].username}, ${users[2].firstName} ${Users.defaultUser.personal.middleName}`,
-      );
-
-      BulkEditSearchPane.clickClearSelectedButton(tastData.usersAccordion);
-      BulkEditSearchPane.verifyClearSelectedButtonExists(tastData.usersAccordion, false);
-      BulkEditSearchPane.verifyLogsPaneHeader();
-
-      BulkEditSearchPane.clickChooseUserUnderUserAccordion();
-      BulkEditSearchPane.fillUserFilterInput(users[1].username);
-      BulkEditSearchPane.verifyDropdown(users[1].username);
-      BulkEditSearchPane.selectUserFromDropdown(users[1].username);
-      BulkEditSearchPane.verifyClearSelectedButtonExists(tastData.usersAccordion);
-      BulkEditSearchPane.resetAllBtnIsDisabled(false);
-      BulkEditSearchPane.verifyCellsValues(
-        8,
-        `${users[1].username}, ${users[1].firstName} ${Users.defaultUser.personal.middleName}`,
-      );
-
-      BulkEditSearchPane.clickChooseUserUnderUserAccordion();
-      BulkEditSearchPane.fillUserFilterInput(users[0].username);
-      BulkEditSearchPane.verifyUserIsNotInUserList(users[0].username);
-      BulkEditSearchPane.verifyEmptyUserDropdown();
-    },
-  );
 });
