+++ cypress/e2e/bulk-edit/logs/in-app/bulk-edit-logs-in-app-holdings-uuids.cy.js
@@ -9,6 +9,7 @@ import getRandomPostfix from '../../../../support/utils/stringTools';
 import BulkEditActions from '../../../../support/fragments/bulk-edit/bulk-edit-actions';
 import BulkEditFiles from '../../../../support/fragments/bulk-edit/bulk-edit-files';
 import InventorySearchAndFilter from '../../../../support/fragments/inventory/inventorySearchAndFilter';
+import BulkEditLogs from '../../../../support/fragments/bulk-edit/bulk-edit-logs';
 
 let user;
 let uuid;
@@ -21,124 +22,128 @@ const item = {
   itemBarcode: getRandomPostfix(),
 };
 
-describe('Bulk Edit - Logs', () => {
-  before('create test data', () => {
-    cy.createTempUser([
-      permissions.bulkEditLogsView.gui,
-      permissions.bulkEditView.gui,
-      permissions.bulkEditEdit.gui,
-      permissions.inventoryAll.gui,
-    ]).then((userProperties) => {
-      user = userProperties;
+describe('bulk-edit', () => {
+  describe('logs', () => {
+    describe('in-app approach', () => {
+      before('create test data', () => {
+        cy.createTempUser([
+          permissions.bulkEditLogsView.gui,
+          permissions.bulkEditView.gui,
+          permissions.bulkEditEdit.gui,
+          permissions.inventoryAll.gui,
+        ]).then((userProperties) => {
+          user = userProperties;
 
-      const instanceId = InventoryInstances.createInstanceViaApi(
-        item.instanceName,
-        item.itemBarcode,
+          const instanceId = InventoryInstances.createInstanceViaApi(
+            item.instanceName,
+            item.itemBarcode,
+          );
+          cy.getHoldings({
+            limit: 1,
+            query: `"instanceId"="${instanceId}"`,
+          }).then((holdings) => {
+            uuid = holdings[0].id;
+            FileManager.createFile(`cypress/fixtures/${validHoldingUUIDsFileName}`, uuid);
+          });
+          cy.login(user.username, user.password, {
+            path: TopMenu.bulkEditPath,
+            waiter: BulkEditSearchPane.waitLoading,
+          });
+        });
+      });
+
+      after('delete test data', () => {
+        cy.getAdminToken();
+        Users.deleteViaApi(user.userId);
+        InventoryInstances.deleteInstanceAndHoldingRecordAndAllItemsViaApi(item.itemBarcode);
+        FileManager.deleteFile(`cypress/fixtures/${validHoldingUUIDsFileName}`);
+        FileManager.deleteFileFromDownloadsByMask(
+          validHoldingUUIDsFileName,
+          `*${matchedRecordsFileNameValid}`,
+          previewOfProposedChangesFileName,
+          updatedRecordsFileName,
+        );
+      });
+
+      it(
+        'C375289 Verify generated Logs files for Holdings In app -- only valid Holdings UUIDs (firebird)',
+        { tags: ['criticalPath', 'firebird'] },
+        () => {
+          BulkEditSearchPane.verifyDragNDropRecordTypeIdentifierArea('Holdings', 'Holdings UUIDs');
+          BulkEditSearchPane.uploadFile(validHoldingUUIDsFileName);
+          BulkEditSearchPane.waitFileUploading();
+
+          BulkEditActions.downloadMatchedResults();
+          BulkEditActions.openInAppStartBulkEditFrom();
+
+          const tempLocation = 'Annex';
+          const permLocation = 'Main Library';
+
+          BulkEditActions.replaceTemporaryLocation(tempLocation, 'holdings', 0);
+          BulkEditActions.addNewBulkEditFilterString();
+          BulkEditActions.replacePermanentLocation(permLocation, 'holdings', 1);
+
+          BulkEditActions.confirmChanges();
+          BulkEditActions.downloadPreview();
+          BulkEditActions.commitChanges();
+          BulkEditSearchPane.waitFileUploading();
+          BulkEditActions.openActions();
+          BulkEditActions.downloadChangedCSV();
+
+          BulkEditSearchPane.openLogsSearch();
+          BulkEditLogs.verifyLogsPane();
+          BulkEditLogs.checkHoldingsCheckbox();
+          BulkEditLogs.clickActionsRunBy(user.username);
+          BulkEditLogs.verifyLogsRowActionWhenCompleted();
+
+          BulkEditLogs.downloadFileUsedToTrigger();
+          BulkEditFiles.verifyCSVFileRows(validHoldingUUIDsFileName, [uuid]);
+
+          BulkEditLogs.downloadFileWithMatchingRecords();
+          BulkEditFiles.verifyMatchedResultFileContent(
+            `*${matchedRecordsFileNameValid}`,
+            [uuid],
+            'firstElement',
+            true,
+          );
+
+          BulkEditLogs.downloadFileWithProposedChanges();
+          BulkEditFiles.verifyMatchedResultFileContent(
+            previewOfProposedChangesFileName,
+            [tempLocation],
+            'temporaryLocation',
+            true,
+          );
+          BulkEditFiles.verifyMatchedResultFileContent(
+            previewOfProposedChangesFileName,
+            [permLocation],
+            'permanentLocation',
+            true,
+          );
+
+          BulkEditLogs.downloadFileWithUpdatedRecords();
+          BulkEditFiles.verifyMatchedResultFileContent(
+            updatedRecordsFileName,
+            [tempLocation],
+            'temporaryLocation',
+            true,
+          );
+          BulkEditFiles.verifyMatchedResultFileContent(
+            updatedRecordsFileName,
+            [permLocation],
+            'permanentLocation',
+            true,
+          );
+
+          cy.visit(TopMenu.inventoryPath);
+          InventorySearchAndFilter.switchToHoldings();
+          InventorySearchAndFilter.searchByParameter('Holdings UUID', uuid);
+          InventorySearchAndFilter.selectSearchResultItem();
+          InventorySearchAndFilter.selectViewHoldings();
+          InventoryInstance.verifyHoldingsPermanentLocation(permLocation);
+          InventoryInstance.verifyHoldingsTemporaryLocation(tempLocation);
+        },
       );
-      cy.getHoldings({
-        limit: 1,
-        query: `"instanceId"="${instanceId}"`,
-      }).then((holdings) => {
-        uuid = holdings[0].id;
-        FileManager.createFile(`cypress/fixtures/${validHoldingUUIDsFileName}`, uuid);
-      });
-      cy.login(user.username, user.password, {
-        path: TopMenu.bulkEditPath,
-        waiter: BulkEditSearchPane.waitLoading,
-      });
     });
   });
-
-  after('delete test data', () => {
-    cy.getAdminToken();
-    Users.deleteViaApi(user.userId);
-    InventoryInstances.deleteInstanceAndHoldingRecordAndAllItemsViaApi(item.itemBarcode);
-    FileManager.deleteFile(`cypress/fixtures/${validHoldingUUIDsFileName}`);
-    FileManager.deleteFileFromDownloadsByMask(
-      validHoldingUUIDsFileName,
-      `*${matchedRecordsFileNameValid}`,
-      previewOfProposedChangesFileName,
-      updatedRecordsFileName,
-    );
-  });
-
-  it(
-    'C375289 Verify generated Logs files for Holdings In app -- only valid Holdings UUIDs (firebird)',
-    { tags: ['criticalPath', 'firebird'] },
-    () => {
-      BulkEditSearchPane.verifyDragNDropRecordTypeIdentifierArea('Holdings', 'Holdings UUIDs');
-      BulkEditSearchPane.uploadFile(validHoldingUUIDsFileName);
-      BulkEditSearchPane.waitFileUploading();
-
-      BulkEditActions.downloadMatchedResults();
-      BulkEditActions.openInAppStartBulkEditFrom();
-
-      const tempLocation = 'Annex';
-      const permLocation = 'Main Library';
-
-      BulkEditActions.replaceTemporaryLocation(tempLocation, 'holdings', 0);
-      BulkEditActions.addNewBulkEditFilterString();
-      BulkEditActions.replacePermanentLocation(permLocation, 'holdings', 1);
-
-      BulkEditActions.confirmChanges();
-      BulkEditActions.downloadPreview();
-      BulkEditActions.commitChanges();
-      BulkEditSearchPane.waitFileUploading();
-      BulkEditActions.openActions();
-      BulkEditActions.downloadChangedCSV();
-
-      BulkEditSearchPane.openLogsSearch();
-      BulkEditSearchPane.verifyLogsPane();
-      BulkEditSearchPane.checkHoldingsCheckbox();
-      BulkEditSearchPane.clickActionsRunBy(user.username);
-      BulkEditSearchPane.verifyLogsRowActionWhenCompleted();
-
-      BulkEditSearchPane.downloadFileUsedToTrigger();
-      BulkEditFiles.verifyCSVFileRows(validHoldingUUIDsFileName, [uuid]);
-
-      BulkEditSearchPane.downloadFileWithMatchingRecords();
-      BulkEditFiles.verifyMatchedResultFileContent(
-        `*${matchedRecordsFileNameValid}`,
-        [uuid],
-        'firstElement',
-        true,
-      );
-
-      BulkEditSearchPane.downloadFileWithProposedChanges();
-      BulkEditFiles.verifyMatchedResultFileContent(
-        previewOfProposedChangesFileName,
-        [tempLocation],
-        'temporaryLocation',
-        true,
-      );
-      BulkEditFiles.verifyMatchedResultFileContent(
-        previewOfProposedChangesFileName,
-        [permLocation],
-        'permanentLocation',
-        true,
-      );
-
-      BulkEditSearchPane.downloadFileWithUpdatedRecords();
-      BulkEditFiles.verifyMatchedResultFileContent(
-        updatedRecordsFileName,
-        [tempLocation],
-        'temporaryLocation',
-        true,
-      );
-      BulkEditFiles.verifyMatchedResultFileContent(
-        updatedRecordsFileName,
-        [permLocation],
-        'permanentLocation',
-        true,
-      );
-
-      cy.visit(TopMenu.inventoryPath);
-      InventorySearchAndFilter.switchToHoldings();
-      InventorySearchAndFilter.searchByParameter('Holdings UUID', uuid);
-      InventorySearchAndFilter.selectSearchResultItem();
-      InventorySearchAndFilter.selectViewHoldings();
-      InventoryInstance.verifyHoldingsPermanentLocation(permLocation);
-      InventoryInstance.verifyHoldingsTemporaryLocation(tempLocation);
-    },
-  );
 });
