+++ cypress/e2e/bulk-edit/logs/in-app/bulk-edit-logs-in-app-items-holdings-uuids.cy.js
@@ -10,6 +10,7 @@ import BulkEditFiles from '../../../../support/fragments/bulk-edit/bulk-edit-fil
 import InventorySearchAndFilter from '../../../../support/fragments/inventory/inventorySearchAndFilter';
 import ItemRecordView from '../../../../support/fragments/inventory/item/itemRecordView';
 import { ITEM_STATUS_NAMES } from '../../../../support/constants';
+import BulkEditLogs from '../../../../support/fragments/bulk-edit/bulk-edit-logs';
 
 let user;
 const validHoldingUUIDsFileName = `validHoldingUUIDs_${getRandomPostfix()}.csv`;
@@ -36,134 +37,143 @@ const inventoryEntity = {
   },
 };
 
-describe('Bulk Edit - Logs', () => {
-  before('create test data', () => {
-    cy.createTempUser([
-      permissions.bulkEditLogsView.gui,
-      permissions.bulkEditView.gui,
-      permissions.bulkEditEdit.gui,
-      permissions.inventoryAll.gui,
-    ]).then((userProperties) => {
-      user = userProperties;
+describe('bulk-edit', () => {
+  describe('logs', () => {
+    describe('in-app approach', () => {
+      before('create test data', () => {
+        cy.createTempUser([
+          permissions.bulkEditLogsView.gui,
+          permissions.bulkEditView.gui,
+          permissions.bulkEditEdit.gui,
+          permissions.inventoryAll.gui,
+        ]).then((userProperties) => {
+          user = userProperties;
 
-      inventoryEntity.instance.id = InventoryInstances.createInstanceViaApi(
-        inventoryEntity.instance.name,
-        inventoryEntity.item.barcode,
+          inventoryEntity.instance.id = InventoryInstances.createInstanceViaApi(
+            inventoryEntity.instance.name,
+            inventoryEntity.item.barcode,
+          );
+          cy.getHoldings({
+            limit: 1,
+            query: `"instanceId"="${inventoryEntity.instance.id}"`,
+          }).then((holdings) => {
+            inventoryEntity.holdingId = holdings[0].id;
+            inventoryEntity.locations.permanent.id = holdings[0].permanentLocationId;
+            FileManager.createFile(
+              `cypress/fixtures/${validHoldingUUIDsFileName}`,
+              inventoryEntity.holdingId,
+            );
+
+            cy.getLocations({
+              limit: 1,
+              query: `id="${inventoryEntity.locations.permanent.id}"`,
+            }).then((loc) => {
+              inventoryEntity.locations.permanent.name = loc.name;
+            });
+            cy.getItems({ query: `"barcode"=="${inventoryEntity.item.barcode}"` }).then(
+              (inventoryItem) => {
+                inventoryEntity.item.id = inventoryItem.id;
+              },
+            );
+            cy.getItems({
+              query: `"barcode"=="secondBarcode_${inventoryEntity.item.barcode}"`,
+            }).then((inventoryItem) => {
+              inventoryEntity.item.id2 = inventoryItem.id;
+            });
+          });
+          cy.login(user.username, user.password, {
+            path: TopMenu.bulkEditPath,
+            waiter: BulkEditSearchPane.waitLoading,
+          });
+        });
+      });
+
+      after('delete test data', () => {
+        cy.getAdminToken();
+        Users.deleteViaApi(user.userId);
+        InventoryInstances.deleteInstanceAndHoldingRecordAndAllItemsViaApi(
+          inventoryEntity.item.barcode,
+        );
+        FileManager.deleteFile(`cypress/fixtures/${validHoldingUUIDsFileName}`);
+        FileManager.deleteFileFromDownloadsByMask(
+          validHoldingUUIDsFileName,
+          `*${matchedRecordsFileNameValid}`,
+          previewOfProposedChangesFileName,
+          updatedRecordsFileName,
+        );
+      });
+
+      it(
+        'C375288 Verify generated Logs files for Items In app -- only valid Holdings UUIDs (firebird)',
+        { tags: ['smoke', 'firebird'] },
+        () => {
+          BulkEditSearchPane.checkItemsRadio();
+          BulkEditSearchPane.selectRecordIdentifier('Holdings UUIDs');
+
+          BulkEditSearchPane.uploadFile(validHoldingUUIDsFileName);
+          BulkEditSearchPane.waitFileUploading();
+          BulkEditActions.downloadMatchedResults();
+          BulkEditActions.openInAppStartBulkEditFrom();
+
+          BulkEditActions.clearTemporaryLocation();
+          BulkEditActions.addNewBulkEditFilterString();
+          BulkEditActions.replacePermanentLocation(
+            inventoryEntity.locations.permanent.name,
+            'item',
+            1,
+          );
+          BulkEditActions.addNewBulkEditFilterString();
+          BulkEditActions.replaceItemStatus(ITEM_STATUS_NAMES.AVAILABLE, 2);
+
+          BulkEditActions.confirmChanges();
+          BulkEditActions.downloadPreview();
+          BulkEditActions.commitChanges();
+          BulkEditSearchPane.waitFileUploading();
+          BulkEditActions.openActions();
+          BulkEditActions.downloadChangedCSV();
+
+          BulkEditSearchPane.openLogsSearch();
+          BulkEditLogs.checkItemsCheckbox();
+          BulkEditLogs.clickActionsRunBy(user.username);
+          BulkEditLogs.verifyLogsRowActionWhenCompleted();
+
+          BulkEditLogs.downloadFileUsedToTrigger();
+          BulkEditFiles.verifyCSVFileRows(validHoldingUUIDsFileName, [inventoryEntity.holdingId]);
+
+          BulkEditLogs.downloadFileWithMatchingRecords();
+          BulkEditFiles.verifyMatchedResultFileContent(
+            `*${matchedRecordsFileNameValid}`,
+            [inventoryEntity.item.id, inventoryEntity.item.id2],
+            'firstElement',
+            true,
+          );
+
+          BulkEditLogs.downloadFileWithProposedChanges();
+          BulkEditFiles.verifyCSVFileRows(previewOfProposedChangesFileName, [
+            '',
+            `${inventoryEntity.locations.permanent.name} > 1,,1,${inventoryEntity.item.barcode}`,
+            `${inventoryEntity.locations.permanent.name} > 1,,1,secondBarcode_${inventoryEntity.item.barcode}`,
+            '',
+          ]);
+
+          BulkEditLogs.downloadFileWithUpdatedRecords();
+          BulkEditFiles.verifyCSVFileRows(updatedRecordsFileName, [
+            '',
+            `${inventoryEntity.locations.permanent.name} > 1,,1,${inventoryEntity.item.barcode}`,
+            `${inventoryEntity.locations.permanent.name} > 1,,1,secondBarcode_${inventoryEntity.item.barcode}`,
+            '',
+          ]);
+
+          cy.visit(TopMenu.inventoryPath);
+          InventorySearchAndFilter.switchToItem();
+          InventorySearchAndFilter.searchByParameter('Barcode', inventoryEntity.item.barcode);
+          ItemRecordView.checkItemDetails(
+            inventoryEntity.locations.permanent.name,
+            inventoryEntity.item.barcode,
+            ITEM_STATUS_NAMES.AVAILABLE,
+          );
+        },
       );
-      cy.getHoldings({
-        limit: 1,
-        query: `"instanceId"="${inventoryEntity.instance.id}"`,
-      }).then((holdings) => {
-        inventoryEntity.holdingId = holdings[0].id;
-        inventoryEntity.locations.permanent.id = holdings[0].permanentLocationId;
-        FileManager.createFile(
-          `cypress/fixtures/${validHoldingUUIDsFileName}`,
-          inventoryEntity.holdingId,
-        );
-
-        cy.getLocations({ limit: 1, query: `id="${inventoryEntity.locations.permanent.id}"` }).then(
-          (loc) => {
-            inventoryEntity.locations.permanent.name = loc.name;
-          },
-        );
-        cy.getItems({ query: `"barcode"=="${inventoryEntity.item.barcode}"` }).then(
-          (inventoryItem) => {
-            inventoryEntity.item.id = inventoryItem.id;
-          },
-        );
-        cy.getItems({ query: `"barcode"=="secondBarcode_${inventoryEntity.item.barcode}"` }).then(
-          (inventoryItem) => {
-            inventoryEntity.item.id2 = inventoryItem.id;
-          },
-        );
-      });
-      cy.login(user.username, user.password, {
-        path: TopMenu.bulkEditPath,
-        waiter: BulkEditSearchPane.waitLoading,
-      });
     });
   });
-
-  after('delete test data', () => {
-    cy.getAdminToken();
-    Users.deleteViaApi(user.userId);
-    InventoryInstances.deleteInstanceAndHoldingRecordAndAllItemsViaApi(
-      inventoryEntity.item.barcode,
-    );
-    FileManager.deleteFile(`cypress/fixtures/${validHoldingUUIDsFileName}`);
-    FileManager.deleteFileFromDownloadsByMask(
-      validHoldingUUIDsFileName,
-      `*${matchedRecordsFileNameValid}`,
-      previewOfProposedChangesFileName,
-      updatedRecordsFileName,
-    );
-  });
-
-  it(
-    'C375288 Verify generated Logs files for Items In app -- only valid Holdings UUIDs (firebird)',
-    { tags: ['smoke', 'firebird'] },
-    () => {
-      BulkEditSearchPane.checkItemsRadio();
-      BulkEditSearchPane.selectRecordIdentifier('Holdings UUIDs');
-
-      BulkEditSearchPane.uploadFile(validHoldingUUIDsFileName);
-      BulkEditSearchPane.waitFileUploading();
-      BulkEditActions.downloadMatchedResults();
-      BulkEditActions.openInAppStartBulkEditFrom();
-
-      BulkEditActions.clearTemporaryLocation();
-      BulkEditActions.addNewBulkEditFilterString();
-      BulkEditActions.replacePermanentLocation(inventoryEntity.locations.permanent.name, 'item', 1);
-      BulkEditActions.addNewBulkEditFilterString();
-      BulkEditActions.replaceItemStatus(ITEM_STATUS_NAMES.AVAILABLE, 2);
-
-      BulkEditActions.confirmChanges();
-      BulkEditActions.downloadPreview();
-      BulkEditActions.commitChanges();
-      BulkEditSearchPane.waitFileUploading();
-      BulkEditActions.openActions();
-      BulkEditActions.downloadChangedCSV();
-
-      BulkEditSearchPane.openLogsSearch();
-      BulkEditSearchPane.checkItemsCheckbox();
-      BulkEditSearchPane.clickActionsRunBy(user.username);
-      BulkEditSearchPane.verifyLogsRowActionWhenCompleted();
-
-      BulkEditSearchPane.downloadFileUsedToTrigger();
-      BulkEditFiles.verifyCSVFileRows(validHoldingUUIDsFileName, [inventoryEntity.holdingId]);
-
-      BulkEditSearchPane.downloadFileWithMatchingRecords();
-      BulkEditFiles.verifyMatchedResultFileContent(
-        `*${matchedRecordsFileNameValid}`,
-        [inventoryEntity.item.id, inventoryEntity.item.id2],
-        'firstElement',
-        true,
-      );
-
-      BulkEditSearchPane.downloadFileWithProposedChanges();
-      BulkEditFiles.verifyCSVFileRows(previewOfProposedChangesFileName, [
-        '',
-        `${inventoryEntity.locations.permanent.name} > 1,,1,${inventoryEntity.item.barcode}`,
-        `${inventoryEntity.locations.permanent.name} > 1,,1,secondBarcode_${inventoryEntity.item.barcode}`,
-        '',
-      ]);
-
-      BulkEditSearchPane.downloadFileWithUpdatedRecords();
-      BulkEditFiles.verifyCSVFileRows(updatedRecordsFileName, [
-        '',
-        `${inventoryEntity.locations.permanent.name} > 1,,1,${inventoryEntity.item.barcode}`,
-        `${inventoryEntity.locations.permanent.name} > 1,,1,secondBarcode_${inventoryEntity.item.barcode}`,
-        '',
-      ]);
-
-      cy.visit(TopMenu.inventoryPath);
-      InventorySearchAndFilter.switchToItem();
-      InventorySearchAndFilter.searchByParameter('Barcode', inventoryEntity.item.barcode);
-      ItemRecordView.checkItemDetails(
-        inventoryEntity.locations.permanent.name,
-        inventoryEntity.item.barcode,
-        ITEM_STATUS_NAMES.AVAILABLE,
-      );
-    },
-  );
 });
