+++ cypress/e2e/bulk-edit/logs/in-app/bulk-edit-logs-in-app-suppressed-holdings-false-items-not.cy.js
@@ -12,6 +12,7 @@ import InventorySearchAndFilter from '../../../../support/fragments/inventory/in
 import ItemRecordView from '../../../../support/fragments/inventory/item/itemRecordView';
 import TopMenuNavigation from '../../../../support/fragments/topMenuNavigation';
 import InstanceRecordView from '../../../../support/fragments/inventory/instanceRecordView';
+import BulkEditLogs from '../../../../support/fragments/bulk-edit/bulk-edit-logs';
 
 // TO DO: remove ignoring errors. Now when you click on one of the buttons, some promise in the application returns false
 Cypress.on('uncaught:exception', () => false);
@@ -26,132 +27,136 @@ const matchedRecordsFileName = `Matched-Records-${instanceHRIDFileName}`;
 const previewOfProposedChangesFileName = `*-Updates-Preview-${instanceHRIDFileName}`;
 const updatedRecordsFileName = `*-Changed-Records*-${instanceHRIDFileName}`;
 
-describe('Bulk Edit - Logs', () => {
-  before('create test data', () => {
-    cy.createTempUser([
-      permissions.inventoryAll.gui,
-      permissions.bulkEditView.gui,
-      permissions.bulkEditEdit.gui,
-      permissions.bulkEditLogsView.gui,
-    ]).then((userProperties) => {
-      user = userProperties;
+describe('bulk-edit', () => {
+  describe('logs', () => {
+    describe('in-app approach', () => {
+      before('create test data', () => {
+        cy.createTempUser([
+          permissions.inventoryAll.gui,
+          permissions.bulkEditView.gui,
+          permissions.bulkEditEdit.gui,
+          permissions.bulkEditLogsView.gui,
+        ]).then((userProperties) => {
+          user = userProperties;
 
-      item.instanceId = InventoryInstances.createInstanceViaApi(
-        item.instanceName,
-        item.itemBarcode,
-      );
-      cy.getHoldings({
-        limit: 1,
-        expandAll: true,
-        query: `"instanceId"="${item.instanceId}"`,
-      }).then((holdings) => {
-        item.holdingsHRID = holdings[0].hrid;
-        cy.updateHoldingRecord(holdings[0].id, {
-          ...holdings[0],
-          discoverySuppress: true,
+          item.instanceId = InventoryInstances.createInstanceViaApi(
+            item.instanceName,
+            item.itemBarcode,
+          );
+          cy.getHoldings({
+            limit: 1,
+            expandAll: true,
+            query: `"instanceId"="${item.instanceId}"`,
+          }).then((holdings) => {
+            item.holdingsHRID = holdings[0].hrid;
+            cy.updateHoldingRecord(holdings[0].id, {
+              ...holdings[0],
+              discoverySuppress: true,
+            });
+          });
+          cy.getInstanceById(item.instanceId).then((body) => {
+            body.discoverySuppress = true;
+            cy.updateInstance(body);
+            item.instanceHRID = body.hrid;
+            FileManager.createFile(`cypress/fixtures/${instanceHRIDFileName}`, item.instanceHRID);
+          });
+          cy.login(user.username, user.password, {
+            path: TopMenu.bulkEditPath,
+            waiter: BulkEditSearchPane.waitLoading,
+          });
         });
       });
-      cy.getInstanceById(item.instanceId).then((body) => {
-        body.discoverySuppress = true;
-        cy.updateInstance(body);
-        item.instanceHRID = body.hrid;
-        FileManager.createFile(`cypress/fixtures/${instanceHRIDFileName}`, item.instanceHRID);
+
+      after('delete test data', () => {
+        cy.getAdminToken();
+        Users.deleteViaApi(user.userId);
+        InventoryInstances.deleteInstanceAndHoldingRecordAndAllItemsViaApi(item.itemBarcode);
+        FileManager.deleteFile(`cypress/fixtures/${instanceHRIDFileName}`);
+        FileManager.deleteFileFromDownloadsByMask(
+          instanceHRIDFileName,
+          `*${matchedRecordsFileName}`,
+          previewOfProposedChangesFileName,
+          updatedRecordsFileName,
+        );
       });
-      cy.login(user.username, user.password, {
-        path: TopMenu.bulkEditPath,
-        waiter: BulkEditSearchPane.waitLoading,
-      });
+
+      it(
+        'C402326 Verify "Suppress from discovery" option is set to False when Holdings are suppressed and Items are not (firebird)',
+        { tags: ['criticalPath', 'firebird'] },
+        () => {
+          BulkEditSearchPane.checkHoldingsRadio();
+          BulkEditSearchPane.selectRecordIdentifier('Instance HRIDs');
+          BulkEditSearchPane.uploadFile(instanceHRIDFileName);
+          BulkEditSearchPane.waitFileUploading();
+          BulkEditSearchPane.verifyMatchedResults(item.holdingsHRID);
+
+          const suppressFromDiscovery = false;
+          BulkEditActions.openActions();
+          BulkEditActions.openInAppStartBulkEditFrom();
+          BulkEditActions.editSuppressFromDiscovery(suppressFromDiscovery, 0, true);
+          BulkEditActions.checkApplyToItemsRecordsCheckbox();
+          BulkEditActions.confirmChanges();
+          BulkEditActions.commitChanges();
+
+          BulkEditSearchPane.waitFileUploading();
+          BulkEditActions.openActions();
+          BulkEditSearchPane.changeShowColumnCheckboxIfNotYet('Suppress from discovery');
+          BulkEditSearchPane.verifyChangesUnderColumns(
+            'Suppress from discovery',
+            suppressFromDiscovery,
+          );
+
+          BulkEditSearchPane.openLogsSearch();
+          BulkEditLogs.checkHoldingsCheckbox();
+          BulkEditLogs.clickActionsRunBy(user.username);
+
+          BulkEditLogs.downloadFileUsedToTrigger();
+          BulkEditFiles.verifyCSVFileRows(instanceHRIDFileName, [item.instanceHRID]);
+
+          BulkEditLogs.downloadFileWithMatchingRecords();
+          BulkEditFiles.verifyMatchedResultFileContent(
+            `*${matchedRecordsFileName}`,
+            [item.instanceHRID],
+            'instanceHrid',
+            true,
+          );
+
+          BulkEditLogs.downloadFileWithProposedChanges();
+          BulkEditFiles.verifyMatchedResultFileContent(
+            previewOfProposedChangesFileName,
+            [item.instanceHRID],
+            'instanceHrid',
+            true,
+          );
+
+          BulkEditLogs.downloadFileWithUpdatedRecords();
+          BulkEditFiles.verifyMatchedResultFileContent(
+            updatedRecordsFileName,
+            [item.instanceHRID],
+            'instanceHrid',
+            true,
+          );
+
+          TopMenuNavigation.navigateToApp('Inventory');
+          InventoryInstances.searchByTitle(item.instanceName);
+          InventoryInstances.selectInstance();
+          InstanceRecordView.verifyMarkAsSuppressedFromDiscovery();
+
+          TopMenuNavigation.navigateToApp('Inventory');
+          InventorySearchAndFilter.switchToItem();
+          InventorySearchAndFilter.searchByParameter('Barcode', item.itemBarcode);
+          ItemRecordView.waitLoading();
+          ItemRecordView.closeDetailView();
+          InventorySearchAndFilter.selectViewHoldings();
+          HoldingsRecordView.checkMarkAsSuppressedFromDiscoveryAbsent();
+
+          TopMenuNavigation.navigateToApp('Inventory');
+          InventorySearchAndFilter.switchToItem();
+          InventorySearchAndFilter.searchByParameter('Barcode', item.itemBarcode);
+          ItemRecordView.waitLoading();
+          ItemRecordView.suppressedAsDiscoveryIsAbsent();
+        },
+      );
     });
   });
-
-  after('delete test data', () => {
-    cy.getAdminToken();
-    Users.deleteViaApi(user.userId);
-    InventoryInstances.deleteInstanceAndHoldingRecordAndAllItemsViaApi(item.itemBarcode);
-    FileManager.deleteFile(`cypress/fixtures/${instanceHRIDFileName}`);
-    FileManager.deleteFileFromDownloadsByMask(
-      instanceHRIDFileName,
-      `*${matchedRecordsFileName}`,
-      previewOfProposedChangesFileName,
-      updatedRecordsFileName,
-    );
-  });
-
-  it(
-    'C402326 Verify "Suppress from discovery" option is set to False when Holdings are suppressed and Items are not (firebird)',
-    { tags: ['criticalPath', 'firebird'] },
-    () => {
-      BulkEditSearchPane.checkHoldingsRadio();
-      BulkEditSearchPane.selectRecordIdentifier('Instance HRIDs');
-      BulkEditSearchPane.uploadFile(instanceHRIDFileName);
-      BulkEditSearchPane.waitFileUploading();
-      BulkEditSearchPane.verifyMatchedResults(item.holdingsHRID);
-
-      const suppressFromDiscovery = false;
-      BulkEditActions.openActions();
-      BulkEditActions.openInAppStartBulkEditFrom();
-      BulkEditActions.editSuppressFromDiscovery(suppressFromDiscovery, 0, true);
-      BulkEditActions.checkApplyToItemsRecordsCheckbox();
-      BulkEditActions.confirmChanges();
-      BulkEditActions.commitChanges();
-
-      BulkEditSearchPane.waitFileUploading();
-      BulkEditActions.openActions();
-      BulkEditSearchPane.changeShowColumnCheckboxIfNotYet('Suppress from discovery');
-      BulkEditSearchPane.verifyChangesUnderColumns(
-        'Suppress from discovery',
-        suppressFromDiscovery,
-      );
-
-      BulkEditSearchPane.openLogsSearch();
-      BulkEditSearchPane.checkHoldingsCheckbox();
-      BulkEditSearchPane.clickActionsRunBy(user.username);
-
-      BulkEditSearchPane.downloadFileUsedToTrigger();
-      BulkEditFiles.verifyCSVFileRows(instanceHRIDFileName, [item.instanceHRID]);
-
-      BulkEditSearchPane.downloadFileWithMatchingRecords();
-      BulkEditFiles.verifyMatchedResultFileContent(
-        `*${matchedRecordsFileName}`,
-        [item.instanceHRID],
-        'instanceHrid',
-        true,
-      );
-
-      BulkEditSearchPane.downloadFileWithProposedChanges();
-      BulkEditFiles.verifyMatchedResultFileContent(
-        previewOfProposedChangesFileName,
-        [item.instanceHRID],
-        'instanceHrid',
-        true,
-      );
-
-      BulkEditSearchPane.downloadFileWithUpdatedRecords();
-      BulkEditFiles.verifyMatchedResultFileContent(
-        updatedRecordsFileName,
-        [item.instanceHRID],
-        'instanceHrid',
-        true,
-      );
-
-      TopMenuNavigation.navigateToApp('Inventory');
-      InventoryInstances.searchByTitle(item.instanceName);
-      InventoryInstances.selectInstance();
-      InstanceRecordView.verifyMarkAsSuppressedFromDiscovery();
-
-      TopMenuNavigation.navigateToApp('Inventory');
-      InventorySearchAndFilter.switchToItem();
-      InventorySearchAndFilter.searchByParameter('Barcode', item.itemBarcode);
-      ItemRecordView.waitLoading();
-      ItemRecordView.closeDetailView();
-      InventorySearchAndFilter.selectViewHoldings();
-      HoldingsRecordView.checkMarkAsSuppressedFromDiscoveryAbsent();
-
-      TopMenuNavigation.navigateToApp('Inventory');
-      InventorySearchAndFilter.switchToItem();
-      InventorySearchAndFilter.searchByParameter('Barcode', item.itemBarcode);
-      ItemRecordView.waitLoading();
-      ItemRecordView.suppressedAsDiscoveryIsAbsent();
-    },
-  );
 });
