+++ cypress/e2e/bulk-edit/logs/in-app/bulk-edit-logs-in-app-suppressed-items.cy.js
@@ -10,6 +10,7 @@ import InventoryItems from '../../../../support/fragments/inventory/item/invento
 import TopMenu from '../../../../support/fragments/topMenu';
 import InventorySearchAndFilter from '../../../../support/fragments/inventory/inventorySearchAndFilter';
 import ItemRecordView from '../../../../support/fragments/inventory/item/itemRecordView';
+import BulkEditLogs from '../../../../support/fragments/bulk-edit/bulk-edit-logs';
 
 let user;
 
@@ -24,110 +25,114 @@ const inventoryEntity = {
   itemId: '',
 };
 
-describe('Bulk Edit - Logs', () => {
-  before('create test data', () => {
-    cy.createTempUser([
-      permissions.bulkEditView.gui,
-      permissions.bulkEditEdit.gui,
-      permissions.bulkEditLogsView.gui,
-      permissions.inventoryAll.gui,
-    ]).then((userProperties) => {
-      user = userProperties;
-      InventoryInstances.createInstanceViaApi(
-        inventoryEntity.instanceName,
-        inventoryEntity.itemBarcode,
-      );
-      cy.getItems({ query: `"barcode"=="${inventoryEntity.itemBarcode}"` }).then(
-        (inventoryItem) => {
-          inventoryItem.discoverySuppress = true;
-          inventoryEntity.itemId = inventoryItem.id;
-          InventoryItems.editItemViaApi(inventoryItem);
+describe('bulk-edit', () => {
+  describe('logs', () => {
+    describe('in-app approach', () => {
+      before('create test data', () => {
+        cy.createTempUser([
+          permissions.bulkEditView.gui,
+          permissions.bulkEditEdit.gui,
+          permissions.bulkEditLogsView.gui,
+          permissions.inventoryAll.gui,
+        ]).then((userProperties) => {
+          user = userProperties;
+          InventoryInstances.createInstanceViaApi(
+            inventoryEntity.instanceName,
+            inventoryEntity.itemBarcode,
+          );
+          cy.getItems({ query: `"barcode"=="${inventoryEntity.itemBarcode}"` }).then(
+            (inventoryItem) => {
+              inventoryItem.discoverySuppress = true;
+              inventoryEntity.itemId = inventoryItem.id;
+              InventoryItems.editItemViaApi(inventoryItem);
+            },
+          );
+          FileManager.createFile(
+            `cypress/fixtures/${validItemBarcodesFileName}`,
+            inventoryEntity.itemBarcode,
+          );
+          cy.login(user.username, user.password, {
+            path: TopMenu.bulkEditPath,
+            waiter: BulkEditSearchPane.waitLoading,
+          });
+        });
+      });
+
+      after('delete test data', () => {
+        cy.getAdminToken();
+        Users.deleteViaApi(user.userId);
+        FileManager.deleteFile(`cypress/fixtures/${validItemBarcodesFileName}`);
+        FileManager.deleteFileFromDownloadsByMask(
+          validItemBarcodesFileName,
+          `*${matchedRecordsFileName}`,
+          previewOfProposedChangesFileName,
+          updatedRecordsFileName,
+        );
+      });
+
+      it(
+        'C380761 Verify generated Logs files for Items suppressed from discovery (firebird)',
+        { tags: ['criticalPath', 'firebird'] },
+        () => {
+          BulkEditSearchPane.checkItemsRadio();
+          BulkEditSearchPane.selectRecordIdentifier('Item barcode');
+
+          BulkEditSearchPane.uploadFile(validItemBarcodesFileName);
+          BulkEditSearchPane.waitFileUploading();
+
+          BulkEditActions.downloadMatchedResults();
+          BulkEditActions.openInAppStartBulkEditFrom();
+          BulkEditActions.editSuppressFromDiscovery(false);
+          BulkEditActions.addNewBulkEditFilterString();
+          BulkEditActions.fillPermanentLoanType('Selected', 1);
+          BulkEditActions.confirmChanges();
+          BulkEditActions.downloadPreview();
+          BulkEditActions.commitChanges();
+          BulkEditSearchPane.waitFileUploading();
+          BulkEditActions.openActions();
+          BulkEditSearchPane.changeShowColumnCheckboxIfNotYet('Suppress from discovery');
+          BulkEditSearchPane.verifyChangesUnderColumns('Suppress from discovery', false);
+          BulkEditActions.downloadChangedCSV();
+
+          BulkEditSearchPane.openLogsSearch();
+          BulkEditLogs.checkItemsCheckbox();
+          BulkEditLogs.clickActionsRunBy(user.username);
+
+          BulkEditLogs.downloadFileUsedToTrigger();
+          BulkEditFiles.verifyCSVFileRows(validItemBarcodesFileName, [inventoryEntity.itemBarcode]);
+
+          BulkEditLogs.downloadFileWithMatchingRecords();
+          BulkEditFiles.verifyMatchedResultFileContent(
+            `*${matchedRecordsFileName}`,
+            [inventoryEntity.itemId],
+            'firstElement',
+            true,
+          );
+
+          BulkEditLogs.downloadFileWithProposedChanges();
+          BulkEditFiles.verifyMatchedResultFileContent(
+            previewOfProposedChangesFileName,
+            [inventoryEntity.itemId],
+            'firstElement',
+            true,
+          );
+
+          BulkEditLogs.downloadFileWithUpdatedRecords();
+          BulkEditFiles.verifyMatchedResultFileContent(
+            updatedRecordsFileName,
+            [inventoryEntity.itemId],
+            'firstElement',
+            true,
+          );
+
+          cy.visit(TopMenu.inventoryPath);
+          InventorySearchAndFilter.switchToItem();
+          InventorySearchAndFilter.searchByParameter('Barcode', inventoryEntity.itemBarcode);
+          ItemRecordView.waitLoading();
+          ItemRecordView.suppressedAsDiscoveryIsAbsent();
+          ItemRecordView.verifyPermanentLoanType('Selected');
         },
       );
-      FileManager.createFile(
-        `cypress/fixtures/${validItemBarcodesFileName}`,
-        inventoryEntity.itemBarcode,
-      );
-      cy.login(user.username, user.password, {
-        path: TopMenu.bulkEditPath,
-        waiter: BulkEditSearchPane.waitLoading,
-      });
     });
   });
-
-  after('delete test data', () => {
-    cy.getAdminToken();
-    Users.deleteViaApi(user.userId);
-    FileManager.deleteFile(`cypress/fixtures/${validItemBarcodesFileName}`);
-    FileManager.deleteFileFromDownloadsByMask(
-      validItemBarcodesFileName,
-      `*${matchedRecordsFileName}`,
-      previewOfProposedChangesFileName,
-      updatedRecordsFileName,
-    );
-  });
-
-  it(
-    'C380761 Verify generated Logs files for Items suppressed from discovery (firebird)',
-    { tags: ['criticalPath', 'firebird'] },
-    () => {
-      BulkEditSearchPane.checkItemsRadio();
-      BulkEditSearchPane.selectRecordIdentifier('Item barcode');
-
-      BulkEditSearchPane.uploadFile(validItemBarcodesFileName);
-      BulkEditSearchPane.waitFileUploading();
-
-      BulkEditActions.downloadMatchedResults();
-      BulkEditActions.openInAppStartBulkEditFrom();
-      BulkEditActions.editSuppressFromDiscovery(false);
-      BulkEditActions.addNewBulkEditFilterString();
-      BulkEditActions.fillPermanentLoanType('Selected', 1);
-      BulkEditActions.confirmChanges();
-      BulkEditActions.downloadPreview();
-      BulkEditActions.commitChanges();
-      BulkEditSearchPane.waitFileUploading();
-      BulkEditActions.openActions();
-      BulkEditSearchPane.changeShowColumnCheckboxIfNotYet('Suppress from discovery');
-      BulkEditSearchPane.verifyChangesUnderColumns('Suppress from discovery', false);
-      BulkEditActions.downloadChangedCSV();
-
-      BulkEditSearchPane.openLogsSearch();
-      BulkEditSearchPane.checkItemsCheckbox();
-      BulkEditSearchPane.clickActionsRunBy(user.username);
-
-      BulkEditSearchPane.downloadFileUsedToTrigger();
-      BulkEditFiles.verifyCSVFileRows(validItemBarcodesFileName, [inventoryEntity.itemBarcode]);
-
-      BulkEditSearchPane.downloadFileWithMatchingRecords();
-      BulkEditFiles.verifyMatchedResultFileContent(
-        `*${matchedRecordsFileName}`,
-        [inventoryEntity.itemId],
-        'firstElement',
-        true,
-      );
-
-      BulkEditSearchPane.downloadFileWithProposedChanges();
-      BulkEditFiles.verifyMatchedResultFileContent(
-        previewOfProposedChangesFileName,
-        [inventoryEntity.itemId],
-        'firstElement',
-        true,
-      );
-
-      BulkEditSearchPane.downloadFileWithUpdatedRecords();
-      BulkEditFiles.verifyMatchedResultFileContent(
-        updatedRecordsFileName,
-        [inventoryEntity.itemId],
-        'firstElement',
-        true,
-      );
-
-      cy.visit(TopMenu.inventoryPath);
-      InventorySearchAndFilter.switchToItem();
-      InventorySearchAndFilter.searchByParameter('Barcode', inventoryEntity.itemBarcode);
-      ItemRecordView.waitLoading();
-      ItemRecordView.suppressedAsDiscoveryIsAbsent();
-      ItemRecordView.verifyPermanentLoanType('Selected');
-    },
-  );
 });
