+++ cypress/e2e/bulk-edit/logs/in-app/bulk-edit-logs-in-app-users-external-ids.cy.js
@@ -8,6 +8,7 @@ import UserEdit from '../../../../support/fragments/users/userEdit';
 import UsersSearchPane from '../../../../support/fragments/users/usersSearchPane';
 import BulkEditActions from '../../../../support/fragments/bulk-edit/bulk-edit-actions';
 import BulkEditFiles from '../../../../support/fragments/bulk-edit/bulk-edit-files';
+import BulkEditLogs from '../../../../support/fragments/bulk-edit/bulk-edit-logs';
 
 let user;
 const externalId = getRandomPostfix();
@@ -16,100 +17,104 @@ const matchedRecordsFileName = `Matched-Records-${userExternalIDsFileName}`;
 const previewOfProposedChangesFileName = `*-Updates-Preview-${userExternalIDsFileName}`;
 const updatedRecordsFileName = `*-Changed-Records*-${userExternalIDsFileName}`;
 
-describe('Bulk Edit - Logs', () => {
-  before('create test data', () => {
-    cy.createTempUser([
-      permissions.bulkEditLogsView.gui,
-      permissions.bulkEditUpdateRecords.gui,
-      permissions.uiUsersView.gui,
-      permissions.uiUserEdit.gui,
-    ]).then((userProperties) => {
-      user = userProperties;
-      cy.login(user.username, user.password, {
-        path: TopMenu.usersPath,
-        waiter: UsersSearchPane.waitLoading,
+describe('bulk-edit', () => {
+  describe('logs', () => {
+    describe('in-app approach', () => {
+      before('create test data', () => {
+        cy.createTempUser([
+          permissions.bulkEditLogsView.gui,
+          permissions.bulkEditUpdateRecords.gui,
+          permissions.uiUsersView.gui,
+          permissions.uiUserEdit.gui,
+        ]).then((userProperties) => {
+          user = userProperties;
+          cy.login(user.username, user.password, {
+            path: TopMenu.usersPath,
+            waiter: UsersSearchPane.waitLoading,
+          });
+          FileManager.createFile(`cypress/fixtures/${userExternalIDsFileName}`, externalId);
+        });
       });
-      FileManager.createFile(`cypress/fixtures/${userExternalIDsFileName}`, externalId);
+
+      after('delete test data', () => {
+        cy.getAdminToken();
+        FileManager.deleteFile(`cypress/fixtures/${userExternalIDsFileName}`);
+        Users.deleteViaApi(user.userId);
+        FileManager.deleteFileFromDownloadsByMask(
+          userExternalIDsFileName,
+          `*${matchedRecordsFileName}`,
+          previewOfProposedChangesFileName,
+          updatedRecordsFileName,
+        );
+      });
+
+      it(
+        'C375247 Verify genetated Logs files for Users In app -- only valid External IDs (firebird)',
+        { tags: ['smoke', 'firebird'] },
+        () => {
+          UsersSearchPane.searchByStatus('Active');
+          UsersSearchPane.searchByUsername(user.username);
+          UserEdit.addExternalId(externalId);
+
+          cy.visit(TopMenu.bulkEditPath);
+          BulkEditSearchPane.waitLoading();
+          BulkEditSearchPane.checkUsersRadio();
+          BulkEditSearchPane.selectRecordIdentifier('External IDs');
+
+          BulkEditSearchPane.uploadFile(userExternalIDsFileName);
+          BulkEditSearchPane.waitFileUploading();
+
+          BulkEditActions.downloadMatchedResults();
+
+          BulkEditActions.openInAppStartBulkEditFrom();
+          BulkEditActions.verifyBulkEditForm();
+          const newEmailDomain = 'google.com';
+          BulkEditActions.replaceEmail('folio.org', newEmailDomain);
+          BulkEditActions.confirmChanges();
+          BulkEditActions.downloadPreview();
+          BulkEditActions.commitChanges();
+          BulkEditSearchPane.waitFileUploading();
+          BulkEditActions.openActions();
+          BulkEditActions.downloadChangedCSV();
+
+          BulkEditSearchPane.openLogsSearch();
+          BulkEditLogs.verifyLogsPane();
+          BulkEditLogs.checkUsersCheckbox();
+          BulkEditLogs.clickActionsRunBy(user.username);
+          BulkEditLogs.verifyLogsRowActionWhenCompleted();
+
+          BulkEditLogs.downloadFileUsedToTrigger();
+          BulkEditFiles.verifyCSVFileRows(userExternalIDsFileName, [externalId]);
+
+          BulkEditLogs.downloadFileWithMatchingRecords();
+          BulkEditFiles.verifyMatchedResultFileContent(
+            `*${matchedRecordsFileName}`,
+            [user.barcode],
+            'userBarcode',
+            true,
+          );
+
+          BulkEditLogs.downloadFileWithProposedChanges();
+          BulkEditFiles.verifyMatchedResultFileContent(
+            previewOfProposedChangesFileName,
+            [newEmailDomain],
+            'emailDomain',
+            true,
+          );
+
+          BulkEditLogs.downloadFileWithUpdatedRecords();
+          BulkEditFiles.verifyMatchedResultFileContent(
+            updatedRecordsFileName,
+            [newEmailDomain],
+            'emailDomain',
+            true,
+          );
+
+          cy.visit(TopMenu.usersPath);
+          UsersSearchPane.searchByUsername(user.username);
+          Users.verifyEmailDomainOnUserDetailsPane(newEmailDomain);
+        },
+      );
     });
   });
-
-  after('delete test data', () => {
-    cy.getAdminToken();
-    FileManager.deleteFile(`cypress/fixtures/${userExternalIDsFileName}`);
-    Users.deleteViaApi(user.userId);
-    FileManager.deleteFileFromDownloadsByMask(
-      userExternalIDsFileName,
-      `*${matchedRecordsFileName}`,
-      previewOfProposedChangesFileName,
-      updatedRecordsFileName,
-    );
-  });
-
-  it(
-    'C375247 Verify genetated Logs files for Users In app -- only valid External IDs (firebird)',
-    { tags: ['smoke', 'firebird'] },
-    () => {
-      UsersSearchPane.searchByStatus('Active');
-      UsersSearchPane.searchByUsername(user.username);
-      UserEdit.addExternalId(externalId);
-
-      cy.visit(TopMenu.bulkEditPath);
-      BulkEditSearchPane.waitLoading();
-      BulkEditSearchPane.checkUsersRadio();
-      BulkEditSearchPane.selectRecordIdentifier('External IDs');
-
-      BulkEditSearchPane.uploadFile(userExternalIDsFileName);
-      BulkEditSearchPane.waitFileUploading();
-
-      BulkEditActions.downloadMatchedResults();
-
-      BulkEditActions.openInAppStartBulkEditFrom();
-      BulkEditActions.verifyBulkEditForm();
-      const newEmailDomain = 'google.com';
-      BulkEditActions.replaceEmail('folio.org', newEmailDomain);
-      BulkEditActions.confirmChanges();
-      BulkEditActions.downloadPreview();
-      BulkEditActions.commitChanges();
-      BulkEditSearchPane.waitFileUploading();
-      BulkEditActions.openActions();
-      BulkEditActions.downloadChangedCSV();
-
-      BulkEditSearchPane.openLogsSearch();
-      BulkEditSearchPane.verifyLogsPane();
-      BulkEditSearchPane.checkUsersCheckbox();
-      BulkEditSearchPane.clickActionsRunBy(user.username);
-      BulkEditSearchPane.verifyLogsRowActionWhenCompleted();
-
-      BulkEditSearchPane.downloadFileUsedToTrigger();
-      BulkEditFiles.verifyCSVFileRows(userExternalIDsFileName, [externalId]);
-
-      BulkEditSearchPane.downloadFileWithMatchingRecords();
-      BulkEditFiles.verifyMatchedResultFileContent(
-        `*${matchedRecordsFileName}`,
-        [user.barcode],
-        'userBarcode',
-        true,
-      );
-
-      BulkEditSearchPane.downloadFileWithProposedChanges();
-      BulkEditFiles.verifyMatchedResultFileContent(
-        previewOfProposedChangesFileName,
-        [newEmailDomain],
-        'emailDomain',
-        true,
-      );
-
-      BulkEditSearchPane.downloadFileWithUpdatedRecords();
-      BulkEditFiles.verifyMatchedResultFileContent(
-        updatedRecordsFileName,
-        [newEmailDomain],
-        'emailDomain',
-        true,
-      );
-
-      cy.visit(TopMenu.usersPath);
-      UsersSearchPane.searchByUsername(user.username);
-      Users.verifyEmailDomainOnUserDetailsPane(newEmailDomain);
-    },
-  );
 });
