+++ cypress/e2e/bulk-edit/logs/in-app/bulk-edit-logs-in-app-verify-selected-filters-persist-logs-tab.cy.js
@@ -6,6 +6,7 @@ import Users from '../../../../support/fragments/users/users';
 import BulkEditActions from '../../../../support/fragments/bulk-edit/bulk-edit-actions';
 import getRandomPostfix from '../../../../support/utils/stringTools';
 import InventoryInstances from '../../../../support/fragments/inventory/inventoryInstances';
+import BulkEditLogs from '../../../../support/fragments/bulk-edit/bulk-edit-logs';
 
 let user;
 const userBarcodesFileName = `userBarcodes_${getRandomPostfix()}.csv`;
@@ -15,83 +16,89 @@ const item = {
   itemBarcode: getRandomPostfix(),
 };
 
-describe('Bulk Edit - Logs', () => {
-  before('Create test data', () => {
-    cy.createTempUser(
-      [
-        Permissions.uiUsersView.gui,
-        Permissions.bulkEditUpdateRecords.gui,
-        Permissions.uiUserEdit.gui,
-        Permissions.inventoryAll.gui,
-        Permissions.bulkEditView.gui,
-        Permissions.bulkEditEdit.gui,
-        Permissions.bulkEditCsvView.gui,
-        Permissions.bulkEditCsvEdit.gui,
-        Permissions.bulkEditLogsView.gui,
-      ],
-      'faculty',
-    ).then((userProperties) => {
-      user = userProperties;
-      cy.login(user.username, user.password, {
-        path: TopMenu.bulkEditPath,
-        waiter: BulkEditSearchPane.waitLoading,
+describe('bulk-edit', () => {
+  describe('logs', () => {
+    describe('in-app approach', () => {
+      before('Create test data', () => {
+        cy.createTempUser(
+          [
+            Permissions.uiUsersView.gui,
+            Permissions.bulkEditUpdateRecords.gui,
+            Permissions.uiUserEdit.gui,
+            Permissions.inventoryAll.gui,
+            Permissions.bulkEditView.gui,
+            Permissions.bulkEditEdit.gui,
+            Permissions.bulkEditCsvView.gui,
+            Permissions.bulkEditCsvEdit.gui,
+            Permissions.bulkEditLogsView.gui,
+          ],
+          'faculty',
+        ).then((userProperties) => {
+          user = userProperties;
+          cy.login(user.username, user.password, {
+            path: TopMenu.bulkEditPath,
+            waiter: BulkEditSearchPane.waitLoading,
+          });
+          InventoryInstances.createInstanceViaApi(item.instanceName, item.itemBarcode);
+          cy.getItems({
+            limit: 1,
+            expandAll: true,
+            query: `"barcode"=="${item.itemBarcode}"`,
+          }).then((res) => {
+            item.itemId = res.id;
+            FileManager.createFile(`cypress/fixtures/${itemUUIDsFileName}`, item.itemId);
+            FileManager.createFile(`cypress/fixtures/${userBarcodesFileName}`, user.barcode);
+          });
+        });
       });
-      InventoryInstances.createInstanceViaApi(item.instanceName, item.itemBarcode);
-      cy.getItems({ limit: 1, expandAll: true, query: `"barcode"=="${item.itemBarcode}"` }).then(
-        (res) => {
-          item.itemId = res.id;
-          FileManager.createFile(`cypress/fixtures/${itemUUIDsFileName}`, item.itemId);
-          FileManager.createFile(`cypress/fixtures/${userBarcodesFileName}`, user.barcode);
+
+      after('delete test data', () => {
+        Users.deleteViaApi(user.userId);
+        FileManager.deleteFile(`cypress/fixtures/${userBarcodesFileName}`);
+        FileManager.deleteFile(`cypress/fixtures/${itemUUIDsFileName}`);
+      });
+
+      it(
+        'C380546 Verify that selected filters persist on Logs tab (firebird) (TaaS)',
+        { tags: ['extendedPath', 'firebird'] },
+        () => {
+          BulkEditSearchPane.verifyDragNDropRecordTypeIdentifierArea('Users', 'User Barcodes');
+
+          BulkEditSearchPane.uploadFile(userBarcodesFileName);
+          BulkEditSearchPane.waitFileUploading();
+          BulkEditSearchPane.verifyMatchedResults(user.barcode);
+
+          BulkEditActions.openActions();
+          BulkEditActions.openInAppStartBulkEditFrom();
+          BulkEditActions.fillPatronGroup('staff (Staff Member)');
+          BulkEditActions.confirmChanges();
+          BulkEditActions.commitChanges();
+          BulkEditSearchPane.verifyChangedResults('staff');
+
+          BulkEditSearchPane.openLogsSearch();
+          BulkEditLogs.verifyLogsPane();
+          BulkEditLogs.checkLogsCheckbox('Data modification');
+          BulkEditLogs.checkHoldingsCheckbox();
+          BulkEditLogs.verifyLogResultsFound();
+
+          BulkEditSearchPane.openIdentifierSearch();
+          BulkEditSearchPane.verifyDragNDropRecordTypeIdentifierArea('Items', 'Item UUIDs');
+          BulkEditSearchPane.uploadFile(itemUUIDsFileName);
+          BulkEditSearchPane.waitFileUploading();
+          const newLocation = 'Online';
+          BulkEditActions.openActions();
+          BulkEditActions.openInAppStartBulkEditFrom();
+          BulkEditActions.replaceTemporaryLocation(newLocation, 'item', 0);
+          BulkEditActions.confirmChanges();
+          BulkEditActions.commitChanges();
+          BulkEditSearchPane.verifyChangedResults('Online');
+
+          BulkEditSearchPane.openLogsSearch();
+          BulkEditLogs.verifyLogResultsFound();
+          BulkEditLogs.verifyCheckboxIsSelected('DATA_MODIFICATION', true);
+          BulkEditLogs.verifyCheckboxIsSelected('HOLDINGS_RECORD', true);
         },
       );
     });
   });
-
-  after('delete test data', () => {
-    Users.deleteViaApi(user.userId);
-    FileManager.deleteFile(`cypress/fixtures/${userBarcodesFileName}`);
-    FileManager.deleteFile(`cypress/fixtures/${itemUUIDsFileName}`);
-  });
-
-  it(
-    'C380546 Verify that selected filters persist on Logs tab (firebird) (TaaS)',
-    { tags: ['extendedPath', 'firebird'] },
-    () => {
-      BulkEditSearchPane.verifyDragNDropRecordTypeIdentifierArea('Users', 'User Barcodes');
-
-      BulkEditSearchPane.uploadFile(userBarcodesFileName);
-      BulkEditSearchPane.waitFileUploading();
-      BulkEditSearchPane.verifyMatchedResults(user.barcode);
-
-      BulkEditActions.openActions();
-      BulkEditActions.openInAppStartBulkEditFrom();
-      BulkEditActions.fillPatronGroup('staff (Staff Member)');
-      BulkEditActions.confirmChanges();
-      BulkEditActions.commitChanges();
-      BulkEditSearchPane.verifyChangedResults('staff');
-
-      BulkEditSearchPane.openLogsSearch();
-      BulkEditSearchPane.verifyLogsPane();
-      BulkEditSearchPane.checkLogsCheckbox('Data modification');
-      BulkEditSearchPane.checkHoldingsCheckbox();
-      BulkEditSearchPane.verifyLogResultsFound();
-
-      BulkEditSearchPane.openIdentifierSearch();
-      BulkEditSearchPane.verifyDragNDropRecordTypeIdentifierArea('Items', 'Item UUIDs');
-      BulkEditSearchPane.uploadFile(itemUUIDsFileName);
-      BulkEditSearchPane.waitFileUploading();
-      const newLocation = 'Online';
-      BulkEditActions.openActions();
-      BulkEditActions.openInAppStartBulkEditFrom();
-      BulkEditActions.replaceTemporaryLocation(newLocation, 'item', 0);
-      BulkEditActions.confirmChanges();
-      BulkEditActions.commitChanges();
-      BulkEditSearchPane.verifyChangedResults('Online');
-
-      BulkEditSearchPane.openLogsSearch();
-      BulkEditSearchPane.verifyLogResultsFound();
-      BulkEditSearchPane.verifyCheckboxIsSelected('DATA_MODIFICATION', true);
-      BulkEditSearchPane.verifyCheckboxIsSelected('HOLDINGS_RECORD', true);
-    },
-  );
 });
