+++ cypress/e2e/bulk-edit/logs/in-app/verify-that-actions-menu-is-hidden-on-the-logs-tab-with-in-app-permissions.cy.js
@@ -5,6 +5,7 @@ import permissions from '../../../../support/dictionary/permissions';
 import getRandomPostfix from '../../../../support/utils/stringTools';
 import FileManager from '../../../../support/utils/fileManager';
 import InventoryInstances from '../../../../support/fragments/inventory/inventoryInstances';
+import BulkEditLogs from '../../../../support/fragments/bulk-edit/bulk-edit-logs';
 
 let user;
 const item = {
@@ -14,59 +15,61 @@ const item = {
 const itemBarcodesFileName = `itemBarcodes_${getRandomPostfix()}.csv`;
 
 describe('bulk-edit', () => {
-  describe('in-app approach', () => {
-    before('create user', () => {
-      cy.createTempUser([
-        permissions.bulkEditView.gui,
-        permissions.bulkEditEdit.gui,
-        permissions.bulkEditLogsView.gui,
-        permissions.inventoryAll.gui,
-        permissions.uiUserEdit.gui,
-        permissions.uiUsersPermissions.gui,
-      ]).then((userProperties) => {
-        user = userProperties;
-        InventoryInstances.createInstanceViaApi(item.instanceName, item.barcode);
-        cy.login(user.username, user.password, {
-          path: TopMenu.bulkEditPath,
-          waiter: BulkEditSearchPane.waitLoading,
+  describe('logs', () => {
+    describe('in-app approach', () => {
+      before('create user', () => {
+        cy.createTempUser([
+          permissions.bulkEditView.gui,
+          permissions.bulkEditEdit.gui,
+          permissions.bulkEditLogsView.gui,
+          permissions.inventoryAll.gui,
+          permissions.uiUserEdit.gui,
+          permissions.uiUsersPermissions.gui,
+        ]).then((userProperties) => {
+          user = userProperties;
+          InventoryInstances.createInstanceViaApi(item.instanceName, item.barcode);
+          cy.login(user.username, user.password, {
+            path: TopMenu.bulkEditPath,
+            waiter: BulkEditSearchPane.waitLoading,
+          });
+          FileManager.createFile(`cypress/fixtures/${itemBarcodesFileName}`, item.barcode);
         });
-        FileManager.createFile(`cypress/fixtures/${itemBarcodesFileName}`, item.barcode);
       });
+
+      after('delete test data', () => {
+        cy.getAdminToken();
+        FileManager.deleteFile(`cypress/fixtures/${itemBarcodesFileName}`);
+        InventoryInstances.deleteInstanceAndHoldingRecordAndAllItemsViaApi(item.barcode);
+        Users.deleteViaApi(user.userId);
+      });
+
+      it(
+        'C367996 Verify that "Actions" menu is hidden on the "Logs" tab with In-app permissions (firebird) (TaaS)',
+        { tags: ['extendedPath', 'firebird'] },
+        () => {
+          BulkEditSearchPane.verifySetCriteriaPaneSpecificTabs('Identifier', 'Logs');
+          BulkEditSearchPane.verifySpecificTabHighlighted('Identifier');
+          BulkEditSearchPane.isHoldingsRadioChecked(false);
+          BulkEditSearchPane.isItemsRadioChecked(false);
+          BulkEditSearchPane.actionsIsAbsent();
+
+          BulkEditSearchPane.checkItemsRadio();
+          BulkEditSearchPane.selectRecordIdentifier('Item barcode');
+          BulkEditSearchPane.uploadFile(itemBarcodesFileName);
+          BulkEditSearchPane.waitFileUploading();
+
+          BulkEditSearchPane.verifyActionsAfterConductedInAppUploading(false);
+
+          BulkEditSearchPane.openLogsSearch();
+          BulkEditLogs.verifyLogsPane();
+          BulkEditSearchPane.actionsIsAbsent();
+
+          BulkEditLogs.checkItemsCheckbox();
+          BulkEditLogs.verifyActionsRunBy(
+            `${user.username}, ${user.firstName} ${Users.defaultUser.personal.middleName}`,
+          );
+        },
+      );
     });
-
-    after('delete test data', () => {
-      cy.getAdminToken();
-      FileManager.deleteFile(`cypress/fixtures/${itemBarcodesFileName}`);
-      InventoryInstances.deleteInstanceAndHoldingRecordAndAllItemsViaApi(item.barcode);
-      Users.deleteViaApi(user.userId);
-    });
-
-    it(
-      'C367996 Verify that "Actions" menu is hidden on the "Logs" tab with In-app permissions (firebird) (TaaS)',
-      { tags: ['extendedPath', 'firebird'] },
-      () => {
-        BulkEditSearchPane.verifySetCriteriaPaneSpecificTabs('Identifier', 'Logs');
-        BulkEditSearchPane.verifySpecificTabHighlighted('Identifier');
-        BulkEditSearchPane.isHoldingsRadioChecked(false);
-        BulkEditSearchPane.isItemsRadioChecked(false);
-        BulkEditSearchPane.actionsIsAbsent();
-
-        BulkEditSearchPane.checkItemsRadio();
-        BulkEditSearchPane.selectRecordIdentifier('Item barcode');
-        BulkEditSearchPane.uploadFile(itemBarcodesFileName);
-        BulkEditSearchPane.waitFileUploading();
-
-        BulkEditSearchPane.verifyActionsAfterConductedInAppUploading(false);
-
-        BulkEditSearchPane.openLogsSearch();
-        BulkEditSearchPane.verifyLogsPane();
-        BulkEditSearchPane.actionsIsAbsent();
-
-        BulkEditSearchPane.checkItemsCheckbox();
-        BulkEditSearchPane.verifyActionsRunBy(
-          `${user.username}, ${user.firstName} ${Users.defaultUser.personal.middleName}`,
-        );
-      },
-    );
   });
 });
