+++ cypress/e2e/bulk-edit/permissions/bulk-edit-landing-page-without-functional-permissions.cy.js
@@ -2,6 +2,7 @@ import TopMenu from '../../../support/fragments/topMenu';
 import permissions from '../../../support/dictionary/permissions';
 import BulkEditSearchPane from '../../../support/fragments/bulk-edit/bulk-edit-search-pane';
 import Users from '../../../support/fragments/users/users';
+import BulkEditLogs from '../../../support/fragments/bulk-edit/bulk-edit-logs';
 
 let user;
 
@@ -46,12 +47,12 @@ describe('bulk-edit', () => {
         BulkEditSearchPane.openLogsSearch();
         BulkEditSearchPane.verifySetCriteriaPaneSpecificTabs('Identifier', 'Logs', 'Query');
         BulkEditSearchPane.verifySpecificTabHighlighted('Logs');
-        BulkEditSearchPane.verifyLogsPane();
-        BulkEditSearchPane.verifyLogsStatusesAccordionExistsAndUnchecked();
-        BulkEditSearchPane.verifyRecordTypesSortedAlphabetically();
-        BulkEditSearchPane.verifyLogsStartedAccordionCollapsed();
-        BulkEditSearchPane.verifyLogsEndedAccordionCollapsed();
-        BulkEditSearchPane.verifyUserAccordionCollapsed();
+        BulkEditLogs.verifyLogsPane();
+        BulkEditLogs.verifyLogsStatusesAccordionExistsAndUnchecked();
+        BulkEditLogs.verifyRecordTypesSortedAlphabetically();
+        BulkEditLogs.verifyLogsStartedAccordionCollapsed();
+        BulkEditLogs.verifyLogsEndedAccordionCollapsed();
+        BulkEditLogs.verifyUserAccordionCollapsed();
       },
     );
   });
