+++ cypress/e2e/bulk-edit/permissions/bulk-edit-with-logs-permissions.cy.js
@@ -2,6 +2,7 @@ import permissions from '../../../support/dictionary/permissions';
 import BulkEditSearchPane from '../../../support/fragments/bulk-edit/bulk-edit-search-pane';
 import TopMenu from '../../../support/fragments/topMenu';
 import Users from '../../../support/fragments/users/users';
+import BulkEditLogs from '../../../support/fragments/bulk-edit/bulk-edit-logs';
 
 let user;
 
@@ -39,12 +40,12 @@ describe('bulk-edit', () => {
         BulkEditSearchPane.verifySetCriteriaPaneSpecificTabs('Identifier');
 
         BulkEditSearchPane.openLogsSearch();
-        BulkEditSearchPane.verifyLogsPane();
-        BulkEditSearchPane.checkHoldingsCheckbox();
-        BulkEditSearchPane.checkUsersCheckbox();
-        BulkEditSearchPane.checkItemsCheckbox();
-        BulkEditSearchPane.clickActionsOnTheRow();
-        BulkEditSearchPane.verifyTriggerLogsAction();
+        BulkEditLogs.verifyLogsPane();
+        BulkEditLogs.checkHoldingsCheckbox();
+        BulkEditLogs.checkUsersCheckbox();
+        BulkEditLogs.checkItemsCheckbox();
+        BulkEditLogs.clickActionsOnTheRow();
+        BulkEditLogs.verifyTriggerLogsAction();
       },
     );
   });
